refactor(T10065): decompose tasks-schema.ts into 8+3 domain files (T9834b — Saga T9831)
Split tasks-schema.ts (2485 LOC, 8 domains, 1190-LOC provenance block) into
schema/ subdirectory with one file per domain + 3 provenance sub-files.
tasks-schema.ts becomes a barrel re-export only. drizzle-kit check passes
with zero DDL diff — zero import-site churn for 169+ consumers.
