+++ packages/core/src/store/schema/provenance/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Provenance graph barrel — re-exports all 3 provenance sub-modules.
+ *
+ * @task T9506 (commits)
+ * @task T9507 (pull-requests)
+ * @task T9508–T9509 (releases)
+ */
+
+export * from './commits.js';
+export * from './pull-requests.js';
+export * from './releases.js';
