+++ packages/core/src/store/tasks-schema.ts
@@ -1,6 +1,9 @@
 /**
  * Drizzle ORM schema for CLEO tasks.db (SQLite via node:sqlite + sqlite-proxy).
  *
+ * This file is a barrel re-export. All table declarations live in subdomain
+ * files under `packages/core/src/store/schema/`.
+ *
  * Tables: tasks, task_dependencies, task_relations, sessions,
  *         session_handoff_entries, task_work_history
  * Archive uses the same tasks table with status = 'archived' + archive metadata.
@@ -9,54 +12,6 @@
  * @task W1-T2
  * @task T1609 session_handoff_entries — write-once handoff table
  */
-
-import type { ArchiveReasonValue, TaskKind, TaskScope, TaskSeverity } from '@cleocode/contracts';
-import {
-  ARCHIVE_REASONS,
-  TASK_KINDS,
-  TASK_RELATION_TYPES,
-  TASK_SCOPES,
-  TASK_SEVERITIES,
-  TASK_SIZES,
-} from '@cleocode/contracts/enums';
-import type { BackgroundJobStatus } from '@cleocode/contracts/jobs';
-import type {
-  BrainReleaseLinkType,
-  CommitConventionalType,
-  CommitFileChangeType,
-  CommitLinkKind,
-  CommitLinkSource,
-  PrLinkKind,
-  PrLinkSource,
-  PrState,
-  ReleaseArtifactType,
-  ReleaseChangeType,
-  ReleaseChannel,
-  ReleaseClassifiedBy,
-  ReleaseImpact,
-  ReleaseKind,
-  ReleaseScheme,
-  ReleaseStatus,
-} from '@cleocode/contracts/provenance';
-import { sql } from 'drizzle-orm';
-import {
-  type AnySQLiteColumn,
-  index,
-  integer,
-  primaryKey,
-  sqliteTable,
-  text,
-  unique,
-} from 'drizzle-orm/sqlite-core';
-import {
-  ADR_STATUSES,
-  GATE_STATUSES,
-  LIFECYCLE_PIPELINE_STATUSES,
-  LIFECYCLE_STAGE_STATUSES,
-  MANIFEST_STATUSES,
-  SESSION_STATUSES,
-  TASK_STATUSES,
-} from './status-registry.js';
 
 export type {
   AgentErrorLogRow,
@@ -82,7 +37,8 @@ export type {
 } from './chain-schema.js';
 // Re-export WarpChain schema tables so drizzle-kit picks them up for migrations.
 export { warpChainInstances, warpChains } from './chain-schema.js';
-
+// Re-export all domain tables, constants, and types from the schema subdirectory.
+export * from './schema/index.js';
 // Re-export status constants and types so existing imports from schema.ts still work.
 export {
   ADR_STATUSES,
@@ -101,2439 +57,3 @@ export {
   TASK_STATUSES,
   type TaskStatus,
 } from './status-registry.js';
-
-// === CANONICAL ENUM CONSTANTS (non-status) ===
-
-/** Task priorities matching DB CHECK constraint on tasks.priority. */
-export const TASK_PRIORITIES = ['critical', 'high', 'medium', 'low'] as const;
-
-/** Task types matching DB CHECK constraint on tasks.type. */
-export const TASK_TYPES = ['epic', 'task', 'subtask'] as const;
-
-/**
- * Union types for the promoted task-axis constants — re-exported here so
- * existing `import { TaskKind, … } from '@cleocode/core/store/tasks-schema'`
- * consumers keep working. Canonical declarations live in
- * `packages/contracts/src/task.ts` (re-routed through `enums.ts` constants).
- */
-export type { TaskKind, TaskScope, TaskSeverity };
-/**
- * Task kind, scope, severity, and size axes — promoted to
- * `@cleocode/contracts/enums` in Phase 0c of the SG-ARCH-SOLID Saga
- * (T9955). Re-exported here so that Drizzle's `text({ enum: ... })`
- * row-type narrowing keeps using the same identifier and every
- * `import * as schema from './tasks-schema.js'` consumer preserves
- * byte-identical access.
- *
- * Docs for each constant live with the canonical declaration in
- * `packages/contracts/src/enums.ts`.
- *
- * @task T944
- * @task T9072
- * @task T9073
- * @task T9955
- */
-export { TASK_KINDS, TASK_SCOPES, TASK_SEVERITIES, TASK_SIZES };
-
-/** Canonical lifecycle stage names matching DB CHECK constraint on lifecycle_stages.stage_name. */
-export const LIFECYCLE_STAGE_NAMES = [
-  'research',
-  'consensus',
-  'architecture_decision',
-  'specification',
-  'decomposition',
-  'implementation',
-  'validation',
-  'testing',
-  'release',
-  'contribution',
-] as const;
-
-/** Gate result values matching DB CHECK constraint on lifecycle_gate_results.result. */
-export const LIFECYCLE_GATE_RESULTS = ['pass', 'fail', 'warn'] as const;
-
-/** Evidence type values matching DB CHECK constraint on lifecycle_evidence.type. */
-export const LIFECYCLE_EVIDENCE_TYPES = ['file', 'url', 'manifest'] as const;
-
-/** Token measurement methods for central token telemetry. */
-export const TOKEN_USAGE_METHODS = ['otel', 'provider_api', 'tokenizer', 'heuristic'] as const;
-
-/** Confidence levels for token measurements. */
-export const TOKEN_USAGE_CONFIDENCE = ['real', 'high', 'estimated', 'coarse'] as const;
-
-/** Transport types for token telemetry. */
-export const TOKEN_USAGE_TRANSPORTS = ['cli', 'api', 'agent', 'unknown'] as const;
-
-/**
- * Task relation types — promoted to `@cleocode/contracts/enums` in
- * Phase 0c of the SG-ARCH-SOLID Saga (T9955). Re-exported here for the
- * Drizzle `text({ enum: TASK_RELATION_TYPES })` row-type narrowing and
- * for existing `import * as schema` consumers.
- */
-export { TASK_RELATION_TYPES };
-
-/** Lifecycle transition types matching DB CHECK constraint on lifecycle_transitions.transition_type. */
-export const LIFECYCLE_TRANSITION_TYPES = ['automatic', 'manual', 'forced'] as const;
-
-/**
- * Truth-grade archive reason values — promoted to
- * `@cleocode/contracts/enums` in Phase 0c of the SG-ARCH-SOLID Saga
- * (T9955). Re-exported here so the Drizzle row type narrows from the
- * same identifier and every existing `import * as schema` consumer
- * keeps working. Full docs live with the canonical declaration in
- * `packages/contracts/src/enums.ts`.
- *
- * @task T1408
- * @epic T1407
- * @task T9955
- */
-export { ARCHIVE_REASONS };
-
-/**
- * Union type for {@link ARCHIVE_REASONS}.
- *
- * T1409: re-exported from `@cleocode/contracts` as the SSoT. The contracts
- * package owns the typed Zod enum (`ArchiveReason`) and the inferred
- * {@link ArchiveReasonValue} union; this alias is kept for back-compat with
- * existing core/store consumers that import `ArchiveReason` from this file.
- */
-export type ArchiveReason = ArchiveReasonValue;
-
-/** External task link types matching DB constraint on external_task_links.link_type. */
-export const EXTERNAL_LINK_TYPES = ['created', 'matched', 'manual', 'transferred'] as const;
-
-/** Sync direction types matching DB constraint on external_task_links.sync_direction. */
-export const SYNC_DIRECTIONS = ['inbound', 'outbound', 'bidirectional'] as const;
-
-// === TASKS TABLE ===
-
-export const tasks = sqliteTable(
-  'tasks',
-  {
-    id: text('id').primaryKey(),
-    title: text('title').notNull(),
-    description: text('description'),
-    status: text('status', {
-      enum: TASK_STATUSES,
-    })
-      .notNull()
-      .default('pending'),
-    priority: text('priority', {
-      enum: TASK_PRIORITIES,
-    })
-      .notNull()
-      .default('medium'),
-    type: text('type', { enum: TASK_TYPES }),
-    /**
-     * Task kind axis — orthogonal to `type`. Defaults to `'work'`.
-     * See {@link TASK_KINDS}. Added by T944. DB column named `role` (T9067 deferral).
-     */
-    kind: text('role', { enum: TASK_KINDS }).notNull().default('work'),
-    /**
-     * Task scope axis — granularity of the work. Defaults to `'feature'`.
-     * See {@link TASK_SCOPES}. Added by T944.
-     */
-    scope: text('scope', { enum: TASK_SCOPES }).notNull().default('feature'),
-    /**
-     * Severity level. Valid for any kind (widened from bug-only by T9073).
-     * DB column 'role' preserved with drizzle alias for kind. OWNER-WRITE-ONLY.
-     * See {@link TASK_SEVERITIES}. Added by T944, widened by T9073.
-     */
-    severity: text('severity', { enum: TASK_SEVERITIES }),
-    parentId: text('parent_id').references((): AnySQLiteColumn => tasks.id, {
-      onDelete: 'set null',
-    }),
-    phase: text('phase'),
-    size: text('size', { enum: TASK_SIZES }),
-    position: integer('position'),
-    positionVersion: integer('position_version').default(0),
-
-    // JSON-serialized complex fields (avoids excessive normalization)
-    labelsJson: text('labels_json').default('[]'),
-    notesJson: text('notes_json').default('[]'),
-    acceptanceJson: text('acceptance_json').default('[]'),
-    filesJson: text('files_json').default('[]'),
-
-    // Provenance — T1899: typed CHECK constraint (production|test-fixture|imported|migrated)
-    origin: text('origin'),
-    blockedBy: text('blocked_by'),
-    epicLifecycle: text('epic_lifecycle'),
-    noAutoComplete: integer('no_auto_complete', { mode: 'boolean' }),
-
-    // Timestamps
-    createdAt: text('created_at').notNull().default(sql`(datetime('now'))`),
-    updatedAt: text('updated_at'),
-    completedAt: text('completed_at'),
-    cancelledAt: text('cancelled_at'),
-    cancellationReason: text('cancellation_reason'),
-
-    // Archive metadata (populated when status = 'archived')
-    archivedAt: text('archived_at'),
-    /**
-     * Truth-grade archive reason. SQL-layer storage remains TEXT; the
-     * 6-value enum (see {@link ARCHIVE_REASONS} / {@link ArchiveReason}) is
-     * enforced by a CHECK constraint applied in migration
-     * `20260424000000_t1408-archive-reason-enum`.
-     *
-     * The drizzle `text()` builder is intentionally left UN-narrowed
-     * (no `$type<ArchiveReason>()`) for now: legacy call-sites in
-     * `task-store.ts`, `sqlite-data-accessor.ts`, `delete.ts`, and the
-     * JSON-to-SQLite migration ingest paths still pass historical strings
-     * (`'completed'`, `'deleted'`, `'migrated'`) that the CHECK rejects.
-     * Tightening the TS type here would break the build before those call
-     * sites are updated; that promotion is tracked as a separate follow-up.
-     * Until then, prefer the {@link ArchiveReason} alias for newly authored
-     * code.
-     *
-     * @task T1408
-     */
-    archiveReason: text('archive_reason'),
-    cycleTimeDays: integer('cycle_time_days'),
-
-    // Verification (JSON-serialized)
-    verificationJson: text('verification_json'),
-
-    // Provenance tracking
-    createdBy: text('created_by'),
-    modifiedBy: text('modified_by'),
-    sessionId: text('session_id').references((): AnySQLiteColumn => sessions.id, {
-      onDelete: 'set null',
-    }),
-    // T060: pipeline stage name (RCASD-IVTR+C). Stored as a plain stage name string.
-    // Not referencing lifecycle_stages.id so that stage binding works without
-    // requiring a lifecycle pipeline record for every task.
-    pipelineStage: text('pipeline_stage'),
-    /** Agent ID that has claimed/is assigned to this task. */
-    assignee: text('assignee'),
-    // IVTR orchestration state — nullable JSON blob. NULL = no loop started. @task T811
-    ivtrState: text('ivtr_state'),
-  },
-  (table) => [
-    index('idx_tasks_status').on(table.status),
-    index('idx_tasks_parent_id').on(table.parentId),
-    index('idx_tasks_phase').on(table.phase),
-    index('idx_tasks_type').on(table.type),
-    index('idx_tasks_priority').on(table.priority),
-    index('idx_tasks_session_id').on(table.sessionId),
-    index('idx_tasks_pipeline_stage').on(table.pipelineStage),
-    index('idx_tasks_assignee').on(table.assignee),
-    // T033 composite indexes
-    index('idx_tasks_parent_status').on(table.parentId, table.status),
-    index('idx_tasks_status_priority').on(table.status, table.priority),
-    index('idx_tasks_type_phase').on(table.type, table.phase),
-    index('idx_tasks_status_archive_reason').on(table.status, table.archiveReason),
-    // T944 kind/scope axes (T9072: TS field renamed role→kind; DB column stays 'role')
-    index('idx_tasks_role').on(table.kind),
-    index('idx_tasks_scope').on(table.scope),
-    index('idx_tasks_role_status').on(table.kind, table.status),
-    // T1126 / T1174 — partial index now schema-expressed (see .where() call below).
-    // The corresponding migration is packages/core/migrations/drizzle-tasks/20260421000002_t1126-sentient-proposal-index/
-    // which remains in place for existing DBs' journal continuity.
-    index('idx_tasks_sentient_proposals_today')
-      .on(sql`date(${table.createdAt})`)
-      .where(sql`${table.labelsJson} LIKE '%sentient-tier2%'`),
-  ],
-);
-
-// === TASK DEPENDENCIES ===
-
-export const taskDependencies = sqliteTable(
-  'task_dependencies',
-  {
-    taskId: text('task_id')
-      .notNull()
-      .references(() => tasks.id, { onDelete: 'cascade' }),
-    dependsOn: text('depends_on')
-      .notNull()
-      .references(() => tasks.id, { onDelete: 'cascade' }),
-  },
-  (table) => [
-    primaryKey({ columns: [table.taskId, table.dependsOn] }),
-    index('idx_deps_depends_on').on(table.dependsOn),
-  ],
-);
-
-// === TASK RELATIONS ===
-
-export const taskRelations = sqliteTable(
-  'task_relations',
-  {
-    taskId: text('task_id')
-      .notNull()
-      .references(() => tasks.id, { onDelete: 'cascade' }),
-    relatedTo: text('related_to')
-      .notNull()
-      .references(() => tasks.id, { onDelete: 'cascade' }),
-    relationType: text('relation_type', {
-      enum: TASK_RELATION_TYPES,
-    })
-      .notNull()
-      .default('related'),
-    reason: text('reason'),
-  },
-  (table) => [
-    primaryKey({ columns: [table.taskId, table.relatedTo] }),
-    index('idx_task_relations_related_to').on(table.relatedTo),
-  ],
-);
-
-// === SESSIONS ===
-
-export const sessions = sqliteTable(
-  'sessions',
-  {
-    id: text('id').primaryKey(),
-    name: text('name').notNull(),
-    status: text('status', {
-      enum: SESSION_STATUSES,
-    })
-      .notNull()
-      .default('active'),
-    scopeJson: text('scope_json').notNull().default('{}'),
-    currentTask: text('current_task').references((): AnySQLiteColumn => tasks.id, {
-      onDelete: 'set null',
-    }),
-    taskStartedAt: text('task_started_at'),
-    agent: text('agent'),
-    notesJson: text('notes_json').default('[]'),
-    tasksCompletedJson: text('tasks_completed_json').default('[]'),
-    tasksCreatedJson: text('tasks_created_json').default('[]'),
-    handoffJson: text('handoff_json'),
-    startedAt: text('started_at').notNull().default(sql`(datetime('now'))`),
-    endedAt: text('ended_at'),
-    // Session chain columns (T4959)
-    previousSessionId: text('previous_session_id').references((): AnySQLiteColumn => sessions.id, {
-      onDelete: 'set null',
-    }),
-    nextSessionId: text('next_session_id').references((): AnySQLiteColumn => sessions.id, {
-      onDelete: 'set null',
-    }),
-    agentIdentifier: text('agent_identifier'),
-    handoffConsumedAt: text('handoff_consumed_at'),
-    handoffConsumedBy: text('handoff_consumed_by'),
-    debriefJson: text('debrief_json'),
-    // Provider adapter tracking (T5240)
-    providerId: text('provider_id'),
-    // Session stats columns (type unification)
-    statsJson: text('stats_json'),
-    resumeCount: integer('resume_count'),
-    gradeMode: integer('grade_mode'),
-    // Owner-auth HMAC token for L4a override authentication (T1118)
-    ownerAuthToken: text('owner_auth_token'),
-    // T9975 — per-agent session isolation columns
-    /** Human-readable agent tag (e.g. "agent-A") for multi-agent isolation. */
-    agentHandle: text('agent_handle'),
-    /** Denormalised scope type ("global" | "epic") — avoids JSON parsing in hot paths. */
-    scopeKind: text('scope_kind'),
-    /** Denormalised scope target ID (e.g. "T9964" for epic sessions). NULL for global. */
-    scopeId: text('scope_id'),
-    /** ISO 8601 timestamp of the last mutation — used by idle auto-end hook. */
-    lastActivity: text('last_activity'),
-  },
-  (table) => [
-    index('idx_sessions_status').on(table.status),
-    index('idx_sessions_previous').on(table.previousSessionId),
-    index('idx_sessions_agent_identifier').on(table.agentIdentifier),
-    index('idx_sessions_started_at').on(table.startedAt),
-    // T033 composite index: getActiveSession hot path
-    index('idx_sessions_status_started_at').on(table.status, table.startedAt),
-    // T9975 — per-agent session isolation indexes
-    index('idx_sessions_agent_handle').on(table.agentHandle),
-    index('idx_sessions_scope_kind_id').on(table.scopeKind, table.scopeId),
-  ],
-);
-
-// === SESSION HANDOFF ENTRIES (T1609 — append-only, write-once) ===
-
-/**
- * Write-once handoff log for sessions.
- *
- * Each session may have AT MOST ONE handoff entry (UNIQUE on session_id).
- * A BEFORE UPDATE trigger in the migration SQL raises ABORT if an update is
- * attempted, making the row physically immutable once inserted.  The
- * `persistHandoff` helper in `packages/core/src/sessions/handoff.ts` is the
- * only authorised writer — callers must use that function, never raw SQL.
- *
- * An AFTER INSERT trigger mirrors the value back to `sessions.handoff_json`
- * so that all existing read paths continue to work without modification.
- *
- * @task T1609
- */
-export const sessionHandoffEntries = sqliteTable(
-  'session_handoff_entries',
-  {
-    /** Auto-increment surrogate key — sessions are addressed by sessionId UNIQUE. */
-    id: integer('id').primaryKey({ autoIncrement: true }),
-    /** FK → sessions.id.  UNIQUE enforces one-handoff-per-session. */
-    sessionId: text('session_id')
-      .notNull()
-      .unique()
-      .references(() => sessions.id, { onDelete: 'cascade' }),
-    /** Serialised HandoffData (or DebriefData) JSON blob. */
-    handoffJson: text('handoff_json').notNull(),
-    /** Wall-clock instant this entry was created. */
-    createdAt: text('created_at').notNull().default(sql`(datetime('now'))`),
-  },
-  (table) => [index('idx_session_handoff_session_id').on(table.sessionId)],
-);
-
-// === TASK WORK HISTORY ===
-
-export const taskWorkHistory = sqliteTable(
-  'task_work_history',
-  {
-    id: integer('id').primaryKey({ autoIncrement: true }),
-    sessionId: text('session_id')
-      .notNull()
-      .references(() => sessions.id, { onDelete: 'cascade' }),
-    taskId: text('task_id')
-      .notNull()
-      .references(() => tasks.id, { onDelete: 'cascade' }),
-    setAt: text('set_at').notNull().default(sql`(datetime('now'))`),
-    clearedAt: text('cleared_at'),
-  },
-  (table) => [index('idx_work_history_session').on(table.sessionId)],
-);
-
-// === LIFECYCLE PIPELINES ===
-
-export const lifecyclePipelines = sqliteTable(
-  'lifecycle_pipelines',
-  {
-    id: text('id').primaryKey(),
-    taskId: text('task_id')
-      .notNull()
-      .references(() => tasks.id, { onDelete: 'cascade' }),
-    status: text('status', {
-      enum: LIFECYCLE_PIPELINE_STATUSES,
-    })
-      .notNull()
-      .default('active'),
-    currentStageId: text('current_stage_id'),
-    startedAt: text('started_at').notNull().default(sql`(datetime('now'))`),
-    completedAt: text('completed_at'),
-    updatedAt: text('updated_at').default(sql`(datetime('now'))`),
-    version: integer('version').notNull().default(1),
-  },
-  (table) => [
-    index('idx_lifecycle_pipelines_task_id').on(table.taskId),
-    index('idx_lifecycle_pipelines_status').on(table.status),
-  ],
-);
-
-// === LIFECYCLE STAGES ===
-
-export const lifecycleStages = sqliteTable(
-  'lifecycle_stages',
-  {
-    id: text('id').primaryKey(),
-    pipelineId: text('pipeline_id')
-      .notNull()
-      .references(() => lifecyclePipelines.id, { onDelete: 'cascade' }),
-    stageName: text('stage_name', { enum: LIFECYCLE_STAGE_NAMES }).notNull(),
-    status: text('status', {
-      enum: LIFECYCLE_STAGE_STATUSES,
-    })
-      .notNull()
-      .default('not_started'),
-    sequence: integer('sequence').notNull(),
-    startedAt: text('started_at'),
-    completedAt: text('completed_at'),
-    blockedAt: text('blocked_at'),
-    blockReason: text('block_reason'),
-    skippedAt: text('skipped_at'),
-    skipReason: text('skip_reason'),
-    notesJson: text('notes_json').default('[]'),
-    metadataJson: text('metadata_json').default('{}'),
-    // RCASD provenance tracking columns (T5100)
-    outputFile: text('output_file'),
-    createdBy: text('created_by'),
-    validatedBy: text('validated_by'),
-    validatedAt: text('validated_at'),
-    validationStatus: text('validation_status', {
-      enum: ['pending', 'in_review', 'approved', 'rejected', 'needs_revision'],
-    }),
-    provenanceChainJson: text('provenance_chain_json'),
-  },
-  (table) => [
-    index('idx_lifecycle_stages_pipeline_id').on(table.pipelineId),
-    index('idx_lifecycle_stages_stage_name').on(table.stageName),
-    index('idx_lifecycle_stages_status').on(table.status),
-    index('idx_lifecycle_stages_validated_by').on(table.validatedBy),
-  ],
-);
-
-// === LIFECYCLE GATE RESULTS ===
-
-export const lifecycleGateResults = sqliteTable(
-  'lifecycle_gate_results',
-  {
-    id: text('id').primaryKey(),
-    stageId: text('stage_id')
-      .notNull()
-      .references(() => lifecycleStages.id, { onDelete: 'cascade' }),
-    gateName: text('gate_name').notNull(),
-    result: text('result', {
-      enum: LIFECYCLE_GATE_RESULTS,
-    }).notNull(),
-    checkedAt: text('checked_at').notNull().default(sql`(datetime('now'))`),
-    checkedBy: text('checked_by').notNull(),
-    details: text('details'),
-    reason: text('reason'),
-  },
-  (table) => [index('idx_lifecycle_gate_results_stage_id').on(table.stageId)],
-);
-
-// === LIFECYCLE EVIDENCE ===
-
-export const lifecycleEvidence = sqliteTable(
-  'lifecycle_evidence',
-  {
-    id: text('id').primaryKey(),
-    stageId: text('stage_id')
-      .notNull()
-      .references(() => lifecycleStages.id, { onDelete: 'cascade' }),
-    uri: text('uri').notNull(),
-    type: text('type', {
-      enum: LIFECYCLE_EVIDENCE_TYPES,
-    }).notNull(),
-    recordedAt: text('recorded_at').notNull().default(sql`(datetime('now'))`),
-    recordedBy: text('recorded_by'),
-    description: text('description'),
-  },
-  (table) => [index('idx_lifecycle_evidence_stage_id').on(table.stageId)],
-);
-
-// === LIFECYCLE TRANSITIONS ===
-
-export const lifecycleTransitions = sqliteTable(
-  'lifecycle_transitions',
-  {
-    id: text('id').primaryKey(),
-    pipelineId: text('pipeline_id')
-      .notNull()
-      .references(() => lifecyclePipelines.id, { onDelete: 'cascade' }),
-    fromStageId: text('from_stage_id')
-      .notNull()
-      .references(() => lifecycleStages.id, { onDelete: 'cascade' }),
-    toStageId: text('to_stage_id')
-      .notNull()
-      .references(() => lifecycleStages.id, { onDelete: 'cascade' }),
-    transitionType: text('transition_type', {
-      enum: LIFECYCLE_TRANSITION_TYPES,
-    })
-      .notNull()
-      .default('automatic'),
-    transitionedBy: text('transitioned_by'),
-    createdAt: text('created_at').notNull().default(sql`(datetime('now'))`),
-  },
-  (table) => [index('idx_lifecycle_transitions_pipeline_id').on(table.pipelineId)],
-);
-
-// === MANIFEST ENTRIES (RCASD provenance — T5100) ===
-
-export const manifestEntries = sqliteTable(
-  'manifest_entries',
-  {
-    id: text('id').primaryKey(),
-    pipelineId: text('pipeline_id').references(() => lifecyclePipelines.id, {
-      onDelete: 'cascade',
-    }),
-    stageId: text('stage_id').references(() => lifecycleStages.id, { onDelete: 'cascade' }),
-    title: text('title').notNull(),
-    date: text('date').notNull(),
-    status: text('status', { enum: MANIFEST_STATUSES }).notNull(),
-    agentType: text('agent_type'),
-    outputFile: text('output_file'),
-    topicsJson: text('topics_json').default('[]'),
-    findingsJson: text('findings_json').default('[]'),
-    linkedTasksJson: text('linked_tasks_json').default('[]'),
-    createdBy: text('created_by'),
-    createdAt: text('created_at').notNull().default(sql`(datetime('now'))`),
-  },
-  (table) => [
-    index('idx_manifest_entries_pipeline_id').on(table.pipelineId),
-    index('idx_manifest_entries_stage_id').on(table.stageId),
-    index('idx_manifest_entries_status').on(table.status),
-  ],
-);
-
-// === PIPELINE MANIFEST (T5581) ===
-
-export const pipelineManifest = sqliteTable(
-  'pipeline_manifest',
-  {
-    id: text('id').primaryKey(),
-    sessionId: text('session_id').references(() => sessions.id, { onDelete: 'set null' }),
-    taskId: text('task_id').references(() => tasks.id, { onDelete: 'set null' }),
-    epicId: text('epic_id').references(() => tasks.id, { onDelete: 'set null' }),
-    type: text('type').notNull(),
-    content: text('content').notNull(),
-    contentHash: text('content_hash'),
-    status: text('status').notNull().default('active'),
-    distilled: integer('distilled', { mode: 'boolean' }).notNull().default(false),
-    brainObsId: text('brain_obs_id'),
-    sourceFile: text('source_file'),
-    metadataJson: text('metadata_json'),
-    createdAt: text('created_at').notNull(),
-    archivedAt: text('archived_at'),
-  },
-  (table) => [
-    index('idx_pipeline_manifest_task_id').on(table.taskId),
-    index('idx_pipeline_manifest_session_id').on(table.sessionId),
-    index('idx_pipeline_manifest_distilled').on(table.distilled),
-    index('idx_pipeline_manifest_status').on(table.status),
-    index('idx_pipeline_manifest_content_hash').on(table.contentHash),
-  ],
-);
-
-// === RELEASE MANIFESTS — REMOVED (T9686-B2) ===
-//
-// The legacy `release_manifests` table (T5580) was merged into the unified
-// `releases` table by migration `20260519010000_t9686b2-unify-releases-tables`.
-// Legacy-migrated rows initially used PK `legacy:<version>`; T9756
-// (migration `20260520163500_t9756-uniform-releases-pk`) rewrote them to
-// the uniform `<projectHash>:<version>` shape used by the new pipeline.
-// Provenance is now discriminated by `tasksJson IS NOT NULL` (legacy) rather
-// than by PK prefix. Legacy status values (`prepared`, `committed`, `tagged`,
-// `pushed`, `rolled_back`) remain admitted by the widened
-// {@link RELEASE_STATUSES} enum.
-//
-// All readers/writers now target the unified `releases` table directly.
-// The `releaseManifests` Drizzle binding has been removed — code that
-// previously imported it must use {@link releases} instead.
-
-// === SCHEMA METADATA ===
-
-export const schemaMeta = sqliteTable('schema_meta', {
-  key: text('key').primaryKey(),
-  value: text('value').notNull(),
-});
-
-// === AUDIT LOG ===
-
-/**
- * Task change audit log — stores every add/update/complete/delete/archive operation.
- * Migrated from legacy JSONL task logs to SQLite per ADR-006/ADR-012.
- * No FK on taskId — log entries must survive task deletion.
- *
- * @task T4837
- */
-export const auditLog = sqliteTable(
-  'audit_log',
-  {
-    id: text('id').primaryKey(),
-    timestamp: text('timestamp').notNull().default(sql`(datetime('now'))`),
-    action: text('action').notNull(),
-    taskId: text('task_id').notNull(),
-    actor: text('actor').notNull().default('system'),
-    detailsJson: text('details_json').default('{}'),
-    beforeJson: text('before_json'),
-    afterJson: text('after_json'),
-    // Dispatch layer columns (migration 20260225200000_audit-log-dispatch-columns)
-    domain: text('domain'),
-    operation: text('operation'),
-    sessionId: text('session_id'),
-    requestId: text('request_id'),
-    durationMs: integer('duration_ms'),
-    success: integer('success'),
-    source: text('source'),
-    gateway: text('gateway'),
-    errorMessage: text('error_message'),
-    // Project correlation (T5334)
-    projectHash: text('project_hash'),
-  },
-  (table) => [
-    index('idx_audit_log_task_id').on(table.taskId),
-    index('idx_audit_log_action').on(table.action),
-    index('idx_audit_log_timestamp').on(table.timestamp),
-    index('idx_audit_log_domain').on(table.domain),
-    index('idx_audit_log_request_id').on(table.requestId),
-    index('idx_audit_log_project_hash').on(table.projectHash),
-    index('idx_audit_log_actor').on(table.actor),
-    // T033 composite indexes
-    index('idx_audit_log_session_timestamp').on(table.sessionId, table.timestamp),
-    index('idx_audit_log_domain_operation').on(table.domain, table.operation),
-  ],
-);
-
-// === TOKEN USAGE ===
-
-/**
- * Central provider-aware token telemetry for CLI and external adapters.
- * Stores measured request/response token counts plus method/confidence metadata.
- */
-export const tokenUsage = sqliteTable(
-  'token_usage',
-  {
-    id: text('id').primaryKey(),
-    createdAt: text('created_at').notNull().default(sql`(datetime('now'))`),
-    provider: text('provider').notNull().default('unknown'),
-    model: text('model'),
-    transport: text('transport', { enum: TOKEN_USAGE_TRANSPORTS }).notNull().default('unknown'),
-    gateway: text('gateway'),
-    domain: text('domain'),
-    operation: text('operation'),
-    sessionId: text('session_id').references(() => sessions.id, { onDelete: 'set null' }),
-    taskId: text('task_id').references(() => tasks.id, { onDelete: 'set null' }),
-    requestId: text('request_id'),
-    inputChars: integer('input_chars').notNull().default(0),
-    outputChars: integer('output_chars').notNull().default(0),
-    inputTokens: integer('input_tokens').notNull().default(0),
-    outputTokens: integer('output_tokens').notNull().default(0),
-    totalTokens: integer('total_tokens').notNull().default(0),
-    method: text('method', { enum: TOKEN_USAGE_METHODS }).notNull().default('heuristic'),
-    confidence: text('confidence', { enum: TOKEN_USAGE_CONFIDENCE }).notNull().default('coarse'),
-    requestHash: text('request_hash'),
-    responseHash: text('response_hash'),
-    metadataJson: text('metadata_json').notNull().default('{}'),
-  },
-  (table) => [
-    index('idx_token_usage_created_at').on(table.createdAt),
-    index('idx_token_usage_request_id').on(table.requestId),
-    index('idx_token_usage_session_id').on(table.sessionId),
-    index('idx_token_usage_task_id').on(table.taskId),
-    index('idx_token_usage_provider').on(table.provider),
-    index('idx_token_usage_transport').on(table.transport),
-    index('idx_token_usage_domain_operation').on(table.domain, table.operation),
-    index('idx_token_usage_method').on(table.method),
-    index('idx_token_usage_gateway').on(table.gateway),
-  ],
-);
-
-// === ARCHITECTURE DECISIONS ===
-
-/**
- * Architecture Decision Records (ADRs) stored in the database.
- * Corresponds to the physical ADR markdown files in .cleo/adrs/.
- * Self-referential FKs (supersedes_id, superseded_by_id, amends_id) are
- * enforced at the DB level by T033 migration.
- */
-export const architectureDecisions = sqliteTable(
-  'architecture_decisions',
-  {
-    id: text('id').primaryKey(),
-    title: text('title').notNull(),
-    status: text('status', { enum: ADR_STATUSES }).notNull().default('proposed'),
-    supersedesId: text('supersedes_id').references(
-      (): AnySQLiteColumn => architectureDecisions.id,
-      { onDelete: 'set null' },
-    ),
-    supersededById: text('superseded_by_id').references(
-      (): AnySQLiteColumn => architectureDecisions.id,
-      { onDelete: 'set null' },
-    ),
-    consensusManifestId: text('consensus_manifest_id').references(
-      (): AnySQLiteColumn => manifestEntries.id,
-      { onDelete: 'set null' },
-    ),
-    content: text('content').notNull(),
-    createdAt: text('created_at').notNull().default(sql`(datetime('now'))`),
-    updatedAt: text('updated_at'),
-    // ADR-017 §5.3 extension columns
-    date: text('date').notNull().default(''),
-    acceptedAt: text('accepted_at'),
-    gate: text('gate', { enum: ['HITL', 'automated'] }),
-    gateStatus: text('gate_status', { enum: GATE_STATUSES }),
-    amendsId: text('amends_id').references((): AnySQLiteColumn => architectureDecisions.id, {
-      onDelete: 'set null',
-    }),
-    filePath: text('file_path').notNull().default(''),
-    // ADR-017 §5.4 cognitive search columns (T4942)
-    summary: text('summary'),
-    keywords: text('keywords'),
-    topics: text('topics'),
-  },
-  (table) => [
-    index('idx_arch_decisions_status').on(table.status),
-    index('idx_arch_decisions_amends_id').on(table.amendsId),
-  ],
-);
-
-// === ADR JUNCTION TABLES (ADR-017 §5.3) ===
-
-/** ADR-to-Task links (soft FK — tasks can be purged) */
-export const adrTaskLinks = sqliteTable(
-  'adr_task_links',
-  {
-    adrId: text('adr_id')
-      .notNull()
-      .references(() => architectureDecisions.id, { onDelete: 'cascade' }),
-    taskId: text('task_id')
-      .notNull()
-      .references(() => tasks.id, { onDelete: 'cascade' }),
-    linkType: text('link_type', {
-      enum: ['related', 'governed_by', 'implements'],
-    })
-      .notNull()
-      .default('related'),
-  },
-  (table) => [
-    primaryKey({ columns: [table.adrId, table.taskId] }),
-    index('idx_adr_task_links_task_id').on(table.taskId),
-  ],
-);
-
-/** ADR cross-reference relationships */
-export const adrRelations = sqliteTable(
-  'adr_relations',
-  {
-    fromAdrId: text('from_adr_id')
-      .notNull()
-      .references(() => architectureDecisions.id, { onDelete: 'cascade' }),
-    toAdrId: text('to_adr_id')
-      .notNull()
-      .references(() => architectureDecisions.id, { onDelete: 'cascade' }),
-    relationType: text('relation_type', {
-      enum: ['supersedes', 'amends', 'related'],
-    }).notNull(),
-  },
-  (table) => [primaryKey({ columns: [table.fromAdrId, table.toAdrId, table.relationType] })],
-);
-
-// === EXTERNAL TASK LINKS (provider-agnostic task reconciliation) ===
-
-/**
- * Tracks links between CLEO tasks and external system tasks (Linear, Jira, GitHub, etc.).
- * Used by the reconciliation engine to match external tasks to existing CLEO tasks,
- * detect updates, and maintain bidirectional traceability.
- *
- * Each row represents one link: one CLEO task ↔ one external task from one provider.
- * A CLEO task MAY have links from multiple providers (e.g., both Linear and GitHub).
- * An external task SHOULD have at most one link per provider.
- */
-export const externalTaskLinks = sqliteTable(
-  'external_task_links',
-  {
-    id: text('id').primaryKey(),
-    taskId: text('task_id')
-      .notNull()
-      .references(() => tasks.id, { onDelete: 'cascade' }),
-    /** Provider identifier (e.g. 'linear', 'jira', 'github', 'gitlab'). */
-    providerId: text('provider_id').notNull(),
-    /** Provider-assigned identifier for the external task (opaque to CLEO). */
-    externalId: text('external_id').notNull(),
-    /** Optional URL to the external task (for human navigation). */
-    externalUrl: text('external_url'),
-    /** Title of the external task at the time of last sync. */
-    externalTitle: text('external_title'),
-    /** How this link was established. */
-    linkType: text('link_type', {
-      enum: EXTERNAL_LINK_TYPES,
-    }).notNull(),
-    /** Direction of the sync that created this link. */
-    syncDirection: text('sync_direction', {
-      enum: SYNC_DIRECTIONS,
-    })
-      .notNull()
-      .default('inbound'),
-    /** Arbitrary provider-specific metadata (JSON). */
-    metadataJson: text('metadata_json').default('{}'),
-    /** When the link was first established. */
-    linkedAt: text('linked_at').notNull().default(sql`(datetime('now'))`),
-    /** When the external task was last synchronized. */
-    lastSyncAt: text('last_sync_at'),
-  },
-  (table) => [
-    index('idx_ext_links_task_id').on(table.taskId),
-    index('idx_ext_links_provider_external').on(table.providerId, table.externalId),
-    index('idx_ext_links_provider_id').on(table.providerId),
-    unique('uq_ext_links_task_provider_external').on(
-      table.taskId,
-      table.providerId,
-      table.externalId,
-    ),
-  ],
-);
-
-// === BACKGROUND JOBS (T641) ===
-
-/**
- * Background job status enum values.
- *
- * - `pending`   – queued but not yet executing
- * - `running`   – actively executing in this process
- * - `complete`  – finished successfully
- * - `failed`    – finished with an error
- * - `cancelled` – explicitly cancelled by a caller
- * - `orphaned`  – was `running` when the process exited; requires human/agent review
- */
-export const BACKGROUND_JOB_STATUSES = [
-  'pending',
-  'running',
-  'complete',
-  'failed',
-  'cancelled',
-  'orphaned',
-] as const;
-
-/**
- * Union type for {@link BACKGROUND_JOB_STATUSES}.
- *
- * Promoted to `@cleocode/contracts/jobs` in Phase 0c of the SG-ARCH-SOLID
- * Saga (T9955). Re-exported here for backward compatibility with every
- * `import { BackgroundJobStatus } from '@cleocode/core/store/tasks-schema'`
- * consumer. The contracts-side declaration and the
- * `(typeof BACKGROUND_JOB_STATUSES)[number]` shape are pinned to remain
- * structurally identical by a compile-time test in
- * `packages/contracts/src/__tests__/jobs.test.ts`.
- */
-export type { BackgroundJobStatus };
-
-/**
- * Durable background job row stored in tasks.db.
- *
- * Jobs survive process restart; any row with `status='running'` at startup
- * is transitioned to `status='orphaned'` so humans/agents can triage them.
- *
- * Timestamps are stored as INTEGER milliseconds-since-epoch to allow
- * range queries without string parsing.
- *
- * @task T641
- */
-export const backgroundJobs = sqliteTable(
-  'background_jobs',
-  {
-    /** Unique job identifier (UUID v4). */
-    id: text('id').primaryKey(),
-    /** Operation name, e.g. "nexus.analyze" or "tasks.sync.reconcile". */
-    operation: text('operation').notNull(),
-    /** Current lifecycle status. */
-    status: text('status', { enum: BACKGROUND_JOB_STATUSES }).notNull().default('pending'),
-    /** When the job was created (ms epoch). */
-    startedAt: integer('started_at').notNull(),
-    /** When the job finished (ms epoch); NULL while running. */
-    completedAt: integer('completed_at'),
-    /** JSON-serialised result payload; NULL on failure or while running. */
-    result: text('result'),
-    /** Human-readable error message; NULL on success or while running. */
-    error: text('error'),
-    /** Execution progress 0-100; NULL until progress is reported. */
-    progress: integer('progress'),
-    /** Last heartbeat timestamp (ms epoch). */
-    heartbeatAt: integer('heartbeat_at').notNull(),
-    /** Agent or session ID that claimed this job; NULL if unclaimed. */
-    claimedBy: text('claimed_by'),
-  },
-  (table) => [
-    index('idx_background_jobs_status').on(table.status),
-    index('idx_background_jobs_operation').on(table.operation),
-    index('idx_background_jobs_claimed_by').on(table.claimedBy),
-    index('idx_background_jobs_started_at').on(table.startedAt),
-  ],
-);
-
-export type BackgroundJobRow = typeof backgroundJobs.$inferSelect;
-export type NewBackgroundJobRow = typeof backgroundJobs.$inferInsert;
-
-// === STATUS REGISTRY (ADR-018) ===
-
-export const statusRegistryTable = sqliteTable(
-  'status_registry',
-  {
-    name: text('name').notNull(),
-    entityType: text('entity_type', {
-      enum: ['task', 'session', 'lifecycle_pipeline', 'lifecycle_stage', 'adr', 'gate', 'manifest'],
-    }).notNull(),
-    namespace: text('namespace', { enum: ['workflow', 'governance', 'manifest'] }).notNull(),
-    description: text('description').notNull(),
-    isTerminal: integer('is_terminal', { mode: 'boolean' }).notNull().default(false),
-  },
-  (table) => [
-    primaryKey({ columns: [table.name, table.entityType] }),
-    index('idx_status_registry_entity_type').on(table.entityType),
-    index('idx_status_registry_namespace').on(table.namespace),
-  ],
-);
-
-export type StatusRegistryRow = typeof statusRegistryTable.$inferSelect;
-
-// === Attachment Storage ===
-
-const ATTACHMENT_OWNER_TYPES = [
-  'task',
-  'observation',
-  'session',
-  'decision',
-  'learning',
-  'pattern',
-] as const;
-
-/**
- * Registry of stored attachments (blob content).
- *
- * Storage path: `.cleo/attachments/sha256/<sha256[0..2]>/<sha256[2..]>.<ext>`
- *
- * @epic T760
- * @task T796
- */
-export const attachments = sqliteTable(
-  'attachments',
-  {
-    /** Unique attachment identifier (UUID v4). */
-    id: text('id').primaryKey(),
-    /** SHA-256 hex digest of the uncompressed content. Unique across the registry. */
-    sha256: text('sha256').notNull().unique(),
-    /** Serialised `Attachment` discriminated union (all kind-specific fields). */
-    attachmentJson: text('attachment_json').notNull(),
-    /** ISO 8601 timestamp when this attachment was first registered. */
-    createdAt: text('created_at').notNull(),
-    /**
-     * How many `attachment_refs` rows point at this attachment.
-     *
-     * When `ref_count` drops to zero the blob is eligible for GC.
-     * Blobs are NEVER auto-deleted — use `cleo docs attachments gc`.
-     */
-    refCount: integer('ref_count').notNull().default(0),
-    /**
-     * Optional human-friendly slug for the attachment, unique per project
-     * (the DB is per-project, so uniqueness is enforced via a partial UNIQUE
-     * INDEX `WHERE slug IS NOT NULL`).
-     *
-     * `cleo docs fetch <slug>` resolves an attachment by slug as a friendlier
-     * alternative to `att_<base62>` IDs or SHA-256 prefixes.
-     *
-     * @task T9636 (Epic T9627 / Saga T9625)
-     */
-    slug: text('slug'),
-    /**
-     * Optional taxonomy classification for the attachment.
-     *
-     * Allowed values:
-     *   - `spec`        — formal specification documents
-     *   - `adr`         — architecture decision records
-     *   - `research`    — research notes and exploration artefacts
-     *   - `handoff`     — session/agent handoff notes
-     *   - `note`        — general notes
-     *   - `llm-readme`  — llms.txt / READMEs intended for LLM consumption
-     *
-     * Enforced at the dispatch layer (not as a DB CHECK constraint) so that
-     * forward-compatible additions don't require a schema migration. The
-     * absence of a value means "unclassified" for backward compatibility.
-     *
-     * @task T9637 (Epic T9627 / Saga T9625)
-     */
-    type: text('type'),
-  },
-  (table) => [index('idx_attachments_sha256').on(table.sha256)],
-);
-
-/**
- * Ref-counted junction table linking attachments to owner entities.
- *
- * A single attachment can be referenced by multiple owners simultaneously
- * (e.g., the same RFC PDF attached to an epic task AND to every child task).
- * Deleting a ref decrements `attachments.ref_count`; when `ref_count` reaches
- * zero the blob file may be purged.
- *
- * @epic T760
- * @task T796
- */
-export const attachmentRefs = sqliteTable(
-  'attachment_refs',
-  {
-    /** ID of the attachment (→ `attachments.id`). */
-    attachmentId: text('attachment_id').notNull(),
-    /**
-     * The domain entity type that owns this ref.
-     *
-     * Restricted to the six supported CLEO entity types.
-     */
-    ownerType: text('owner_type', { enum: ATTACHMENT_OWNER_TYPES }).notNull(),
-    /** The ID of the owning entity (e.g., `"T766"`, `"O-abc123"`, `"ses_..."`). */
-    ownerId: text('owner_id').notNull(),
-    /** ISO 8601 timestamp when this ref was created. */
-    attachedAt: text('attached_at').notNull(),
-    /**
-     * Agent identity (or `"human"`) that created this ref.
-     *
-     * Optional; populated from the active session credential when available.
-     */
-    attachedBy: text('attached_by'),
-  },
-  (table) => [
-    primaryKey({ columns: [table.attachmentId, table.ownerType, table.ownerId] }),
-    index('idx_attachment_refs_attachment_id').on(table.attachmentId),
-    index('idx_attachment_refs_owner').on(table.ownerType, table.ownerId),
-  ],
-);
-
-// === EXPERIMENTS SIDE-TABLE (T944) ===
-
-/**
- * Experiment metadata side-table, keyed 1:1 to `tasks.id` for rows where
- * `tasks.role='experiment'`.
- *
- * Tracks sandbox/branch state and metrics delta for experimental work so the
- * main `tasks` table stays clean and sparse. Cascades on task deletion.
- *
- * @task T944
- */
-export const experiments = sqliteTable(
-  'experiments',
-  {
-    /** Owning task ID. Primary key + cascade FK to {@link tasks}. */
-    taskId: text('task_id')
-      .primaryKey()
-      .references(() => tasks.id, { onDelete: 'cascade' }),
-    /** Git branch used as the experiment sandbox (nullable until created). */
-    sandboxBranch: text('sandbox_branch'),
-    /** Baseline commit SHA the experiment forked from. */
-    baselineCommit: text('baseline_commit'),
-    /** ISO 8601 timestamp when the experiment merged back (null = open). */
-    mergedAt: text('merged_at'),
-    /** Optional receipt ID linking to an audit/receipt record. */
-    receiptId: text('receipt_id'),
-    /** JSON-serialised metrics delta between baseline and experiment. */
-    metricsDeltaJson: text('metrics_delta_json'),
-    /** ISO 8601 timestamp of row creation. */
-    createdAt: text('created_at')
-      .notNull()
-      .$defaultFn(() => new Date().toISOString()),
-    /** ISO 8601 timestamp of last update. */
-    updatedAt: text('updated_at')
-      .notNull()
-      .$defaultFn(() => new Date().toISOString()),
-  },
-  (table) => [index('idx_experiments_merged').on(table.mergedAt)],
-);
-
-// === PROVENANCE GRAPH TABLES (T9506 — ADR-073 / SPEC-T9345) ===
-
-/**
- * Canonical enum values for {@link commits.conventionalType}.
- *
- * Mirrors the Conventional Commits specification prefixes plus `breaking`
- * to flag BREAKING CHANGE footers. Stored as TEXT in SQLite.
- *
- * @task T9506
- */
-export const COMMIT_CONVENTIONAL_TYPES = [
-  'feat',
-  'fix',
-  'chore',
-  'docs',
-  'refactor',
-  'test',
-  'build',
-  'ci',
-  'perf',
-  'revert',
-  'breaking',
-] as const;
-
-/**
- * Union type for {@link COMMIT_CONVENTIONAL_TYPES}. Promoted to
- * `@cleocode/contracts/provenance` in Phase 0c (T9955); re-exported here
- * for backward compatibility. Structurally pinned to the `as const`
- * array above via `packages/contracts/src/__tests__/provenance.test.ts`.
- */
-export type { CommitConventionalType };
-
-/**
- * Link-kind enum for {@link taskCommits.linkKind}.
- *
- * Describes the semantic relationship between a task and a commit:
- * - `implements` — the commit directly implements the task's acceptance criteria
- * - `fixes`      — the commit fixes a bug or regression in the task's work
- * - `refactors`  — the commit restructures code introduced by the task
- * - `tests`      — the commit adds or updates tests for the task
- * - `docs`       — the commit updates documentation for the task
- * - `reverts`    — the commit reverts work introduced by the task
- *
- * @task T9506
- */
-export const COMMIT_LINK_KINDS = [
-  'implements',
-  'fixes',
-  'refactors',
-  'tests',
-  'docs',
-  'reverts',
-] as const;
-
-/**
- * Union type for {@link COMMIT_LINK_KINDS}. Promoted to
- * `@cleocode/contracts/provenance` in Phase 0c (T9955); re-exported here
- * for backward compatibility.
- */
-export type { CommitLinkKind };
-
-/**
- * Link-source enum for {@link taskCommits.linkSource}.
- *
- * Describes how a task↔commit link was discovered:
- * - `commit-trailer` — extracted from a `T####:` or `Task-Id:` git trailer
- * - `commit-subject` — matched `T####` regex in the commit subject line
- * - `pr-title`       — matched task ID in the PR title
- * - `pr-body`        — matched task ID in the PR body markdown
- * - `branch-name`    — parsed from the branch name (e.g., `feat/T9506-...`)
- * - `manual`         — explicitly linked via `cleo provenance link`
- *
- * @task T9506
- */
-export const COMMIT_LINK_SOURCES = [
-  'commit-trailer',
-  'commit-subject',
-  'pr-title',
-  'pr-body',
-  'branch-name',
-  'manual',
-] as const;
-
-/**
- * Union type for {@link COMMIT_LINK_SOURCES}. Promoted to
- * `@cleocode/contracts/provenance` in Phase 0c (T9955); re-exported here
- * for backward compatibility.
- */
-export type { CommitLinkSource };
-
-/**
- * Change-type enum for {@link commitFiles.changeType}.
- *
- * Uses git status letter codes: A=added, M=modified, D=deleted, R=renamed, C=copied.
- *
- * @task T9506
- */
-export const COMMIT_FILE_CHANGE_TYPES = ['A', 'M', 'D', 'R', 'C'] as const;
-
-/**
- * Union type for {@link COMMIT_FILE_CHANGE_TYPES}. Promoted to
- * `@cleocode/contracts/provenance` in Phase 0c (T9955); re-exported here
- * for backward compatibility.
- */
-export type { CommitFileChangeType };
-
-/**
- * `commits` — Every git commit reachable from a release tag.
- *
- * Captures commit identity, author attribution, Conventional Commits
- * classification, and project correlation so that SQL queries can answer
- * "which commits shipped in release vX.Y.Z?" without shelling out to git
- * at query time.
- *
- * Enabled edges (once junction tables are populated):
- *   commit → task     via `task_commits`
- *   commit → release  via `release_commits` (T9507)
- *   commit → file     via `commit_files`
- *
- * @task T9506
- * @epic T9491
- * @see SPEC-T9345 §3.1
- */
-export const commits = sqliteTable(
-  'commits',
-  {
-    /** Full 40-char git SHA — primary key. */
-    sha: text('sha').primaryKey(),
-    /** First 7 characters of the SHA for display purposes. */
-    shortSha: text('short_sha').notNull(),
-    /** Author display name (from `git log %an`). */
-    authorName: text('author_name'),
-    /** Author email (from `git log %ae`). */
-    authorEmail: text('author_email'),
-    /** ISO-8601 author timestamp (from `git log %aI`). */
-    authoredAt: text('authored_at').notNull(),
-    /** Committer display name (from `git log %cn`). */
-    committerName: text('committer_name'),
-    /** Committer email (from `git log %ce`). */
-    committerEmail: text('committer_email'),
-    /** ISO-8601 committer timestamp (from `git log %cI`). */
-    committedAt: text('committed_at').notNull(),
-    /** Full commit message body (subject + blank line + body if present). */
-    message: text('message').notNull(),
-    /** First line of the commit message (the subject). */
-    subject: text('subject').notNull(),
-    /**
-     * Conventional Commits type parsed from the subject line.
-     * NULL when the commit does not follow Conventional Commits format.
-     * See {@link COMMIT_CONVENTIONAL_TYPES}.
-     */
-    conventionalType: text('conventional_type'),
-    /** 1 when this commit matches the `chore(release): vX.Y.Z` release pattern. */
-    isReleaseCommit: integer('is_release_commit').notNull().default(0),
-    /** 1 when the commit has more than one parent (merge commit). */
-    isMergeCommit: integer('is_merge_commit').notNull().default(0),
-    /** JSON array of parent SHA strings. Most commits have exactly one element. */
-    parentShas: text('parent_shas').notNull().default('[]'),
-    /** 0=signature absent or invalid, 1=signature verified, NULL=not checked. */
-    signatureVerified: integer('signature_verified'),
-    /** Best-effort: branch HEAD was on at commit time. NULL if unavailable. */
-    branchAtCommit: text('branch_at_commit'),
-    /**
-     * Project hash from `audit_log.project_hash` — correlates commits to a
-     * specific CLEO project in multi-repo installs.
-     */
-    projectHash: text('project_hash'),
-    /** ISO-8601 timestamp when this row was inserted into tasks.db. */
-    createdAt: text('created_at').notNull().default(sql`(datetime('now'))`),
-  },
-  (table) => [
-    index('idx_commits_short_sha').on(table.shortSha),
-    index('idx_commits_author_email').on(table.authorEmail),
-    index('idx_commits_authored_at').on(table.authoredAt),
-    index('idx_commits_conventional_type').on(table.conventionalType),
-    index('idx_commits_is_release').on(table.isReleaseCommit),
-    index('idx_commits_project_hash').on(table.projectHash),
-  ],
-);
-
-/**
- * `task_commits` — M:N junction between tasks and commits.
- *
- * Each row records one typed linkage edge (task ↔ commit × link_kind).
- * The composite PK `(task_id, commit_sha, link_kind)` permits a single commit
- * to be linked to a single task via multiple relationship types (e.g., both
- * `implements` and `tests`).
- *
- * `task_id` uses ON DELETE SET NULL so provenance history is preserved even
- * when a task is purged from the system.
- * `commit_sha` uses ON DELETE CASCADE — a purged commit removes its junction rows.
- *
- * @task T9506
- * @epic T9491
- * @see SPEC-T9345 §3.2
- */
-export const taskCommits = sqliteTable(
-  'task_commits',
-  {
-    /**
-     * FK → tasks.id. Uses ON DELETE SET NULL to preserve linkage history
-     * after task deletion. NULL = orphaned link (task was purged).
-     */
-    taskId: text('task_id').references(() => tasks.id, { onDelete: 'set null' }),
-    /** FK → commits.sha. ON DELETE CASCADE — purging a commit clears its links. */
-    commitSha: text('commit_sha')
-      .notNull()
-      .references(() => commits.sha, { onDelete: 'cascade' }),
-    /**
-     * Semantic classification of the task↔commit relationship.
-     * See {@link COMMIT_LINK_KINDS}.
-     */
-    linkKind: text('link_kind').notNull(),
-    /**
-     * How this link was discovered (commit-trailer, branch-name, manual, etc.).
-     * See {@link COMMIT_LINK_SOURCES}.
-     */
-    linkSource: text('link_source').notNull(),
-    /** ISO-8601 timestamp when this link was created. */
-    createdAt: text('created_at').notNull().default(sql`(datetime('now'))`),
-  },
-  (table) => [
-    primaryKey({ columns: [table.taskId, table.commitSha, table.linkKind] }),
-    index('idx_task_commits_task_id').on(table.taskId),
-    index('idx_task_commits_commit_sha').on(table.commitSha),
-    index('idx_task_commits_link_kind').on(table.linkKind),
-  ],
-);
-
-/**
- * `commit_files` — Per-file × SHA materialization (blast-radius enabler).
- *
- * Every `(commit_sha, path)` pair has one row describing the type of change
- * (A/M/D/R/C), diff stats, and whether the file is binary. Enables queries
- * such as "which tasks last modified this file path?" and diff-by-file for
- * `cleo release diff`.
- *
- * `old_path` is non-NULL only for R (rename) and C (copy) change types.
- * FK uses ON DELETE CASCADE — purging a commit removes its file rows.
- *
- * @task T9506
- * @epic T9491
- * @see SPEC-T9345 §3.3
- */
-export const commitFiles = sqliteTable(
-  'commit_files',
-  {
-    /** FK → commits.sha. ON DELETE CASCADE. */
-    commitSha: text('commit_sha')
-      .notNull()
-      .references(() => commits.sha, { onDelete: 'cascade' }),
-    /** Canonical repo-relative file path (after rename, if any). */
-    path: text('path').notNull(),
-    /**
-     * Previous path before a rename or copy operation.
-     * NULL for A / M / D change types.
-     */
-    oldPath: text('old_path'),
-    /**
-     * Git status letter: A=added, M=modified, D=deleted, R=renamed, C=copied.
-     * See {@link COMMIT_FILE_CHANGE_TYPES}.
-     */
-    changeType: text('change_type').notNull(),
-    /** Lines added in this commit for this file (0 for deletions and binary files). */
-    linesAdded: integer('lines_added').notNull().default(0),
-    /** Lines deleted in this commit for this file (0 for additions and binary files). */
-    linesDeleted: integer('lines_deleted').notNull().default(0),
-    /** 1 if the file is binary (diff stats are 0/0). */
-    isBinary: integer('is_binary').notNull().default(0),
-  },
-  (table) => [
-    primaryKey({ columns: [table.commitSha, table.path] }),
-    index('idx_commit_files_path').on(table.path),
-    index('idx_commit_files_change_type').on(table.changeType),
-  ],
-);
-
-// ---------------------------------------------------------------------------
-// T9507 — PR provenance graph tables (provenance graph 2/4)
-// Refs: SPEC-T9345 §3.4 (pull_requests), §3.5 (pr_commits, pr_tasks)
-// ---------------------------------------------------------------------------
-
-/**
- * State enum for {@link pullRequests.state}.
- *
- * Mirrors the GitHub PR state machine:
- * - `open`   — PR is open and not yet merged
- * - `closed` — PR was closed without merging
- * - `merged` — PR was merged into the target branch
- *
- * @task T9507
- */
-export const PR_STATES = ['open', 'closed', 'merged'] as const;
-
-/**
- * Union type for {@link PR_STATES}. Promoted to
- * `@cleocode/contracts/provenance` in Phase 0c (T9955); re-exported here
- * for backward compatibility.
- */
-export type { PrState };
-
-/**
- * Link-source enum for {@link prTasks.linkSource}.
- *
- * Describes how a PR↔task link was discovered:
- * - `pr-title`       — task ID matched in the PR title
- * - `pr-body`        — task ID matched in the PR body markdown
- * - `branch-name`    — parsed from the branch name (e.g. `feat/T9507-...`)
- * - `commit-trailer` — extracted from a git trailer in a PR commit
- * - `manual`         — explicitly linked via `cleo provenance link`
- *
- * @task T9507
- */
-export const PR_LINK_SOURCES = [
-  'pr-title',
-  'pr-body',
-  'branch-name',
-  'commit-trailer',
-  'manual',
-] as const;
-
-/**
- * Union type for {@link PR_LINK_SOURCES}. Promoted to
- * `@cleocode/contracts/provenance` in Phase 0c (T9955); re-exported here
- * for backward compatibility.
- */
-export type { PrLinkSource };
-
-/**
- * Link-kind enum for {@link prTasks.linkKind}.
- *
- * Extends {@link COMMIT_LINK_KINDS} with `'tracks'` for PRs that observe
- * a task without directly implementing, fixing, or documenting it.
- *
- * - `implements` — the PR directly implements the task's acceptance criteria
- * - `fixes`      — the PR fixes a bug or regression in the task's work
- * - `refactors`  — the PR restructures code introduced by the task
- * - `tests`      — the PR adds or updates tests for the task
- * - `docs`       — the PR updates documentation for the task
- * - `reverts`    — the PR reverts work introduced by the task
- * - `tracks`     — the PR is related to but does not directly implement the task
- *
- * @task T9507
- */
-export const PR_LINK_KINDS = [
-  'implements',
-  'fixes',
-  'refactors',
-  'tests',
-  'docs',
-  'reverts',
-  'tracks',
-] as const;
-
-/**
- * Union type for {@link PR_LINK_KINDS}. Promoted to
- * `@cleocode/contracts/provenance` in Phase 0c (T9955); re-exported here
- * for backward compatibility.
- */
-export type { PrLinkKind };
-
-/**
- * `pull_requests` — PR metadata for the provenance graph.
- *
- * Captures PR identity, lifecycle state, and release classification so that
- * SQL queries can answer "which PRs shipped in vX.Y.Z?" and "which tasks did
- * PR #185 implement?" without calling the GitHub API at query time.
- *
- * ID format: `"<projectHash>:<prNumber>"` — unique per CLEO project,
- * avoids collision in multi-repo installs.
- *
- * `head_sha` and `merge_commit_sha` are soft FKs to `commits.sha` (ON DELETE
- * SET NULL) so that a commit purge does not cascade-delete PR records.
- *
- * @task T9507
- * @epic T9491
- * @see SPEC-T9345 §3.4
- */
-export const pullRequests = sqliteTable(
-  'pull_requests',
-  {
-    /** Primary key: `"<projectHash>:<prNumber>"`. */
-    id: text('id').primaryKey(),
-    /** GitHub PR number (unique within a repo). */
-    prNumber: integer('pr_number').notNull(),
-    /** Full URL of the repository (e.g. `https://github.com/org/repo`). */
-    repoUrl: text('repo_url').notNull(),
-    /** PR title. */
-    title: text('title').notNull(),
-    /** PR body markdown (may be NULL if the PR had no description). */
-    body: text('body'),
-    /**
-     * Current PR state. See {@link PR_STATES}.
-     * Values: `open` | `closed` | `merged`.
-     */
-    state: text('state').notNull(),
-    /** Target branch name (e.g. `main`). */
-    baseRef: text('base_ref').notNull(),
-    /** Source branch name (e.g. `release/v2026.5.74`). */
-    headRef: text('head_ref').notNull(),
-    /**
-     * HEAD SHA of the PR at merge time.
-     * Soft FK → commits.sha (ON DELETE SET NULL).
-     */
-    headSha: text('head_sha').references(() => commits.sha, { onDelete: 'set null' }),
-    /**
-     * The actual merge commit SHA (NULL if the PR was not merged).
-     * Soft FK → commits.sha (ON DELETE SET NULL).
-     */
-    mergeCommitSha: text('merge_commit_sha').references(() => commits.sha, {
-      onDelete: 'set null',
-    }),
-    /** GitHub login of the PR author. */
-    authorLogin: text('author_login'),
-    /** ISO-8601 timestamp when the PR was opened. */
-    openedAt: text('opened_at').notNull(),
-    /** ISO-8601 timestamp when the PR was merged (NULL if not merged). */
-    mergedAt: text('merged_at'),
-    /** ISO-8601 timestamp when the PR was closed (NULL if still open). */
-    closedAt: text('closed_at'),
-    /** 1 if this PR is a release PR (e.g. opened by `cleo release ship`). */
-    isReleasePr: integer('is_release_pr').notNull().default(0),
-    /** CalVer version string this PR ships (NULL for non-release PRs). */
-    releaseVersion: text('release_version'),
-    /** 1 if this PR contains only a version bump (no feature/fix commits). */
-    isBumpOnly: integer('is_bump_only').notNull().default(0),
-    /**
-     * Project hash from `audit_log.project_hash` — correlates PRs to a
-     * specific CLEO project in multi-repo installs.
-     */
-    projectHash: text('project_hash'),
-    /** ISO-8601 timestamp when this row was first inserted into tasks.db. */
-    createdAt: text('created_at').notNull().default(sql`(datetime('now'))`),
-    /** ISO-8601 timestamp of the last update to this row. */
-    updatedAt: text('updated_at').notNull().default(sql`(datetime('now'))`),
-  },
-  (table) => [
-    index('idx_pr_number').on(table.prNumber),
-    index('idx_pr_state').on(table.state),
-    index('idx_pr_merge_commit_sha').on(table.mergeCommitSha),
-    index('idx_pr_head_sha').on(table.headSha),
-    index('idx_pr_release_version').on(table.releaseVersion),
-    index('idx_pr_project_hash').on(table.projectHash),
-  ],
-);
-
-/**
- * `pr_commits` — M:N ordered junction between pull_requests and commits.
- *
- * Each row records one commit that was part of a PR, along with its ordinal
- * position within the PR commit list. The composite PK `(pr_id, commit_sha)`
- * ensures a commit appears at most once per PR.
- *
- * Both FKs use ON DELETE CASCADE — purging a PR or commit removes its junction rows.
- *
- * @task T9507
- * @epic T9491
- * @see SPEC-T9345 §3.5
- */
-export const prCommits = sqliteTable(
-  'pr_commits',
-  {
-    /** FK → pull_requests.id. ON DELETE CASCADE. */
-    prId: text('pr_id')
-      .notNull()
-      .references(() => pullRequests.id, { onDelete: 'cascade' }),
-    /** FK → commits.sha. ON DELETE CASCADE. */
-    commitSha: text('commit_sha')
-      .notNull()
-      .references(() => commits.sha, { onDelete: 'cascade' }),
-    /** Ordinal position of this commit within the PR commit list (0-based). */
-    position: integer('position').notNull(),
-  },
-  (table) => [
-    primaryKey({ columns: [table.prId, table.commitSha] }),
-    index('idx_pr_commits_pr_id').on(table.prId),
-    index('idx_pr_commits_commit_sha').on(table.commitSha),
-    index('idx_pr_commits_position').on(table.prId, table.position),
-  ],
-);
-
-/**
- * `pr_tasks` — M:N junction between pull_requests and tasks.
- *
- * Each row records one typed link between a PR and a CLEO task. The composite
- * PK `(pr_id, task_id, link_kind)` allows a PR to be linked to the same task
- * via multiple relationship types (e.g., both `implements` and `fixes`).
- *
- * `pr_id` uses ON DELETE CASCADE — if a PR is purged its task links go too.
- * `task_id` uses ON DELETE SET NULL — if a task is purged the PR audit trail
- * is preserved as an orphaned link row (task_id = NULL).
- *
- * @task T9507
- * @epic T9491
- * @see SPEC-T9345 §3.5
- */
-export const prTasks = sqliteTable(
-  'pr_tasks',
-  {
-    /** FK → pull_requests.id. ON DELETE CASCADE. */
-    prId: text('pr_id')
-      .notNull()
-      .references(() => pullRequests.id, { onDelete: 'cascade' }),
-    /**
-     * FK → tasks.id. ON DELETE SET NULL to preserve the PR audit trail
-     * after task deletion. NULL = orphaned link (task was purged).
-     */
-    taskId: text('task_id').references(() => tasks.id, { onDelete: 'set null' }),
-    /**
-     * How this link was discovered. See {@link PR_LINK_SOURCES}.
-     */
-    linkSource: text('link_source').notNull(),
-    /**
-     * Semantic relationship classification. See {@link PR_LINK_KINDS}.
-     */
-    linkKind: text('link_kind').notNull(),
-    /** ISO-8601 timestamp when this link was created. */
-    createdAt: text('created_at').notNull().default(sql`(datetime('now'))`),
-  },
-  (table) => [
-    primaryKey({ columns: [table.prId, table.taskId, table.linkKind] }),
-    index('idx_pr_tasks_pr_id').on(table.prId),
-    index('idx_pr_tasks_task_id').on(table.taskId),
-    index('idx_pr_tasks_link_source').on(table.linkSource),
-  ],
-);
-
-// === PROVENANCE GRAPH TABLES — RELEASES (T9508 — ADR-073 / SPEC-T9345) ===
-
-/**
- * Versioning scheme enum for {@link releases.scheme}.
- *
- * - `calver`        — YYYY.MM.patch (e.g. 2026.5.74) — CLEO default
- * - `semver`        — MAJOR.MINOR.PATCH (e.g. 1.2.3)
- * - `calver-suffix` — YYYY.MM.patch.N suffix hotfix (e.g. 2026.5.74.2)
- *
- * @task T9508
- */
-export const RELEASE_SCHEMES = ['calver', 'semver', 'calver-suffix'] as const;
-
-/**
- * Union type for {@link RELEASE_SCHEMES}. Promoted to
- * `@cleocode/contracts/provenance` in Phase 0c (T9955); re-exported here
- * for backward compatibility.
- */
-export type { ReleaseScheme };
-
-/**
- * Release channel enum for {@link releases.channel}.
- *
- * Controls which npm dist-tag (or equivalent for non-npm archetypes) the
- * artifact is published under. Resolved from git branch via
- * `release-config.ts:resolveChannelFromBranch`.
- *
- * @task T9508
- */
-export const RELEASE_CHANNELS = ['latest', 'beta', 'dev', 'hotfix'] as const;
-
-/**
- * Union type for {@link RELEASE_CHANNELS}. Promoted to
- * `@cleocode/contracts/provenance` in Phase 0c (T9955); re-exported here
- * for backward compatibility.
- */
-export type { ReleaseChannel };
-
-/**
- * Release kind enum for {@link releases.releaseKind}.
- *
- * Describes the whole release packaging, orthogonal to individual
- * {@link RELEASE_CHANGE_TYPES} within the release.
- * - `regular`    — standard scheduled release
- * - `hotfix`     — emergency patch outside regular cadence
- * - `prerelease` — alpha/beta/rc release for early adopters
- *
- * @task T9508
- */
-export const RELEASE_KINDS = ['regular', 'hotfix', 'prerelease'] as const;
-
-/**
- * Union type for {@link RELEASE_KINDS}. Promoted to
- * `@cleocode/contracts/provenance` in Phase 0c (T9955); re-exported here
- * for backward compatibility.
- */
-export type { ReleaseKind };
-
-/**
- * Release status FSM enum for {@link releases.status}.
- *
- * The unified `releases` table (T9686-B2) admits values from both the
- * new T9492 pipeline and the legacy T5580 pipeline. The status value itself
- * discriminates which pipeline owns the row — no separate `pipeline` column
- * is needed.
- *
- * **New T9492 pipeline** — SPEC-T9345 §10.1 FSM:
- *   `planned → pr-opened → pr-merged → published → reconciled`
- *   with `rolled_back | failed | cancelled` as terminal off-ramps.
- *
- * **Legacy T5580 pipeline** — pre-T9492 12-step flow:
- *   `prepared → committed → tagged → pushed`
- *   with `rolled_back` as terminal off-ramp. (No `draft` — never used in live data.)
- *
- * State transitions per R-302 MUST be monotonic within each lifecycle;
- * illegal transitions return `E_INVALID_STATE` and MUST NOT mutate any row.
- *
- * @task T9508
- * @task T9686 (lifecycle union)
- * @see SPEC-T9345 §10.1
- */
-export const RELEASE_STATUSES = [
-  // New T9492 pipeline statuses
-  'planned',
-  'pr-opened',
-  'pr-merged',
-  'published',
-  'reconciled',
-  // Legacy T5580 pipeline statuses (merged in by T9686-B2)
-  'prepared',
-  'committed',
-  'tagged',
-  'pushed',
-  // Shared terminal states
-  'rolled_back',
-  'failed',
-  'cancelled',
-] as const;
-
-/**
- * Union type for {@link RELEASE_STATUSES}. Promoted to
- * `@cleocode/contracts/provenance` in Phase 0c (T9955); re-exported here
- * for backward compatibility. Admits BOTH new T9492 pipeline statuses
- * AND legacy T5580 pipeline statuses (T9686 unification).
- */
-export type { ReleaseStatus };
-
-/**
- * Release change type enum for {@link releaseChanges.changeType}.
- *
- * 12-value CLEO taxonomy (Option B from provenance-graph-design.md §2.2).
- * Lives at the release-changes level (not on `tasks.kind`) so that:
- * - A single task can produce multiple change rows across releases.
- * - Hotfix classification is a release-packaging decision, not a task property.
- * - Auto-classification is agent-writable without touching OWNER-WRITE-ONLY fields.
- *
- * @task T9508
- * @see SPEC-T9345 §2.2
- */
-export const RELEASE_CHANGE_TYPES = [
-  'feature',
-  'enhancement',
-  'bug',
-  'hotfix',
-  'security',
-  'breaking',
-  'refactor',
-  'docs',
-  'chore',
-  'revert',
-  'deprecation',
-  'infrastructure',
-] as const;
-
-/**
- * Union type for {@link RELEASE_CHANGE_TYPES}. Promoted to
- * `@cleocode/contracts/provenance` in Phase 0c (T9955); re-exported here
- * for backward compatibility.
- */
-export type { ReleaseChangeType };
-
-/**
- * Impact level enum for {@link releaseChanges.impact}.
- *
- * Maps to semver bump assessment:
- * - `major` — breaking change (MAJOR bump)
- * - `minor` — new feature (MINOR bump)
- * - `patch` — bug fix / chore (PATCH bump)
- * - `none`  — cosmetic / docs / trivial (no version bump warranted alone)
- *
- * @task T9508
- */
-export const RELEASE_IMPACTS = ['major', 'minor', 'patch', 'none'] as const;
-
-/**
- * Union type for {@link RELEASE_IMPACTS}. Promoted to
- * `@cleocode/contracts/provenance` in Phase 0c (T9955); re-exported here
- * for backward compatibility.
- */
-export type { ReleaseImpact };
-
-/**
- * Classification provenance enum for {@link releaseChanges.classifiedBy}.
- *
- * Tracks how a change was classified:
- * - `auto`     — derived by the classification engine from CC prefix + heuristics
- * - `manual`   — owner overrode the auto classification via `cleo release classify`
- * - `approved` — owner approved an agent-proposed classification
- *
- * @task T9508
- */
-export const RELEASE_CLASSIFIED_BY = ['auto', 'manual', 'approved'] as const;
-
-/**
- * Union type for {@link RELEASE_CLASSIFIED_BY}. Promoted to
- * `@cleocode/contracts/provenance` in Phase 0c (T9955); re-exported here
- * for backward compatibility.
- */
-export type { ReleaseClassifiedBy };
-
-/**
- * `releases` — Canonical release record (ADR-073 / SPEC-T9345 §3.6).
- *
- * As of T9686-B2, this is the SINGLE source of truth for release state. The
- * legacy `release_manifests` table (T5580) was merged into this table and
- * dropped. Legacy-migrated rows initially used PK `legacy:<version>`;
- * T9756 (migration `20260520163500_t9756-uniform-releases-pk`) rewrote
- * them to the uniform `<projectHash>:<version>` shape. All status values
- * (`prepared`/`committed`/`tagged`/`pushed`/`rolled_back` for legacy,
- * `planned`/`pr-opened`/`pr-merged`/`published`/`reconciled` for new)
- * are admitted by the widened {@link RELEASE_STATUSES} enum.
- *
- * ID format (uniform post-T9756):
- *   `<projectHash>:<version>` (e.g. `abc123:v2026.6.0`)
- *
- * Provenance discriminator: `tasksJson IS NOT NULL` indicates a
- * legacy-origin row; new-pipeline rows leave `tasksJson` NULL and store
- * per-task changes in {@link releaseChanges}.
- *
- * Status FSM (lifecycle is discriminated by status value, not a separate column):
- * - New T9492: `planned → pr-opened → pr-merged → published → reconciled`
- * - Legacy T5580: `prepared → committed → tagged → pushed`
- * - Shared terminals: `rolled_back | failed | cancelled`
- *
- * See {@link RELEASE_STATUSES} and SPEC-T9345 §10.1.
- *
- * @task T9508
- * @task T9686 (unification — legacy columns + widened status enum)
- * @task T9756 (uniform PK shape)
- * @epic T9491
- * @see SPEC-T9345 §3.6
- */
-export const releases = sqliteTable(
-  'releases',
-  {
-    /**
-     * Canonical PK — uniform `<projectHash>:<version>` shape (post-T9756).
-     * Example: `1e3146b7352b:v2026.6.0`.
-     *
-     * Historically (T9686-B2 era) legacy-migrated rows used `legacy:<version>`;
-     * the T9756 migration rewrites those to the uniform shape. Provenance is
-     * now signalled by `tasksJson` (non-null = legacy origin), not by PK
-     * prefix.
-     */
-    id: text('id').primaryKey(),
-    /** Release version string, e.g. `v2026.6.0`. UNIQUE — one row per version. */
-    version: text('version').notNull().unique(),
-    /**
-     * Versioning scheme used for this release. See {@link RELEASE_SCHEMES}.
-     * Defaulted to `calver` on legacy migrated rows.
-     */
-    scheme: text('scheme', { enum: RELEASE_SCHEMES }).notNull().default('calver'),
-    /**
-     * Publication channel governing the npm dist-tag (or equivalent).
-     * See {@link RELEASE_CHANNELS}. Defaulted to `latest` on legacy migrated rows.
-     */
-    channel: text('channel', { enum: RELEASE_CHANNELS }).notNull().default('latest'),
-    /**
-     * FK → tasks.id (ON DELETE SET NULL). The epic that scoped this release.
-     * NULL for hotfixes scoped to a single task, not an epic.
-     */
-    epicId: text('epic_id').references(() => tasks.id, { onDelete: 'set null' }),
-    /**
-     * Release packaging kind — describes the whole release, not individual changes.
-     * See {@link RELEASE_KINDS}. Defaulted to `regular` on legacy migrated rows.
-     */
-    releaseKind: text('release_kind', { enum: RELEASE_KINDS }).notNull().default('regular'),
-    /**
-     * Current FSM status. Admits both new-pipeline and legacy-pipeline values
-     * post-T9686-B2 — see {@link RELEASE_STATUSES}.
-     */
-    status: text('status', { enum: RELEASE_STATUSES }).notNull().default('planned'),
-    /** Previous release version string (denormalized for fast prior-release walks). */
-    previousVersion: text('previous_version'),
-    /**
-     * Git merge commit SHA. For new-pipeline rows this is the commit that
-     * landed the release branch into main (populated by release-prepare.yml
-     * at pr-merged time). For legacy-migrated rows this is the ship commit
-     * from the legacy `release_manifests.commit_sha` column.
-     *
-     * HARD FK to `commits(sha)` ON DELETE SET NULL (T9755): T9686-B2 had
-     * dropped this FK because legacy ship SHAs were not present in the
-     * `commits` table; T9755 backfilled rows for every v5.80–v5.88 ship
-     * and PR-merge commit referenced by `releases`, then rebuilt this
-     * table with the FK restored. New ship rows continue to satisfy the
-     * FK because release-publish writes `commits` rows during the
-     * provenance reconcile step before inserting the `releases` row.
-     */
-    mergeCommitSha: text('merge_commit_sha').references(() => commits.sha, {
-      onDelete: 'set null',
-    }),
-    /**
-     * FK → pull_requests.id (ON DELETE SET NULL). The bump-PR opened by `cleo release open`.
-     * Always NULL on legacy migrated rows (legacy pipeline pre-dated `pull_requests`).
-     */
-    prId: text('pr_id'),
-    /** URL of the GitHub Actions workflow run that built and published this release. */
-    workflowRunUrl: text('workflow_run_url'),
-    /** ISO-8601 timestamp when this row was inserted. */
-    createdAt: text('created_at').notNull().default(sql`(datetime('now'))`),
-    /** ISO-8601 timestamp when `cleo release plan` created the plan (new pipeline). */
-    plannedAt: text('planned_at'),
-    /** ISO-8601 timestamp when the bump-PR was opened (new pipeline). */
-    prOpenedAt: text('pr_opened_at'),
-    /** ISO-8601 timestamp when the bump-PR was merged (new pipeline). */
-    prMergedAt: text('pr_merged_at'),
-    /** ISO-8601 timestamp when npm/cargo/etc. publish completed (new pipeline). */
-    publishedAt: text('published_at'),
-    /** ISO-8601 timestamp when `cleo release reconcile` completed (new pipeline). */
-    reconciledAt: text('reconciled_at'),
-    /** ISO-8601 timestamp when `cleo release rollback` completed (terminal). */
-    rolledBackAt: text('rolled_back_at'),
-    /** ISO-8601 timestamp when a failure was detected (terminal). */
-    failedAt: text('failed_at'),
-    /** ISO-8601 timestamp when the operator cancelled the release (terminal). */
-    cancelledAt: text('cancelled_at'),
-    /** Human-readable reason for the failure (populated when status='failed'). */
-    failureReason: text('failure_reason'),
-    /** Agent or operator identity that initiated the rollback. */
-    rolledBackBy: text('rolled_back_by'),
-    /** Project hash for multi-repo CLEO installs (matches audit_log.project_hash). */
-    projectHash: text('project_hash'),
-    // ── Legacy-only columns (merged in by T9686-B2 from `release_manifests`) ──
-    /**
-     * Legacy: JSON array of task IDs included in this release. Populated for
-     * legacy-migrated rows only; NULL on new-pipeline rows (new pipeline
-     * tracks per-task changes in {@link releaseChanges} instead).
-     */
-    tasksJson: text('tasks_json'),
-    /**
-     * Legacy: free-form CHANGELOG body text. Populated for legacy-migrated
-     * rows only; NULL on new-pipeline rows (new pipeline derives the
-     * CHANGELOG from `release_changes` + `release_commits`).
-     */
-    changelog: text('changelog'),
-    /** Legacy: free-form release notes. Populated for legacy-migrated rows only. */
-    notes: text('notes'),
-    /** Legacy: git tag string. Populated for legacy-migrated rows only. */
-    gitTag: text('git_tag'),
-    /** Legacy: ISO-8601 timestamp when the release was marked `prepared`. */
-    preparedAt: text('prepared_at'),
-    /** Legacy: ISO-8601 timestamp when the release was marked `committed`. */
-    committedAt: text('committed_at'),
-    /** Legacy: ISO-8601 timestamp when the release was marked `tagged`. */
-    taggedAt: text('tagged_at'),
-    /** Legacy: ISO-8601 timestamp when the release was marked `pushed` (published). */
-    pushedAt: text('pushed_at'),
-  },
-  (table) => [
-    index('idx_releases_version').on(table.version),
-    index('idx_releases_status').on(table.status),
-    index('idx_releases_channel').on(table.channel),
-    index('idx_releases_epic_id').on(table.epicId),
-    index('idx_releases_merge_commit_sha').on(table.mergeCommitSha),
-    index('idx_releases_project_hash').on(table.projectHash),
-    index('idx_releases_published_at').on(table.publishedAt),
-    index('idx_releases_pushed_at').on(table.pushedAt),
-  ],
-);
-
-/**
- * `release_commits` — M:N junction between releases and commits (SPEC-T9345 §3.8).
- *
- * Each row records one commit that belongs to a given release range, derived from
- * `git log <prev_tag>..<tag>` at ship time and persisted for query speed.
- *
- * `position` is topo-sorted ascending: 0 = oldest commit, N = newest/tag commit.
- *
- * Mutual-exclusivity constraint (application-layer enforced):
- * `is_first`, `is_last`, and `is_release_chore` are logically mutually exclusive —
- * a commit should hold at most one of these roles. Enforced by parity tests and
- * application-layer validation rather than a SQLite CHECK because the three-way
- * `CHECK (is_first + is_last + is_release_chore <= 1)` would require an ALTER TABLE
- * for older SQLite versions that don't support partial CHECK syntax cleanly.
- *
- * FKs:
- *   release_id → releases(id) ON DELETE CASCADE
- *   commit_sha → commits(sha) ON DELETE CASCADE
- *
- * @task T9508
- * @epic T9491
- * @see SPEC-T9345 §3.8
- */
-export const releaseCommits = sqliteTable(
-  'release_commits',
-  {
-    /** FK → releases.id. ON DELETE CASCADE — purging a release removes its junction rows. */
-    releaseId: text('release_id')
-      .notNull()
-      .references(() => releases.id, { onDelete: 'cascade' }),
-    /** FK → commits.sha. ON DELETE CASCADE — purging a commit removes its junction rows. */
-    commitSha: text('commit_sha')
-      .notNull()
-      .references(() => commits.sha, { onDelete: 'cascade' }),
-    /** Topo-sorted ascending position: 0 = oldest commit reachable since prev release. */
-    position: integer('position').notNull(),
-    /**
-     * 1 if this is the first commit after the previous release boundary.
-     * MUTUALLY EXCLUSIVE with is_last and is_release_chore (application enforced).
-     */
-    isFirst: integer('is_first').notNull().default(0),
-    /**
-     * 1 if this is the tag/merge commit that closed this release.
-     * MUTUALLY EXCLUSIVE with is_first and is_release_chore (application enforced).
-     */
-    isLast: integer('is_last').notNull().default(0),
-    /**
-     * 1 for "chore(release): vX.Y.Z" version-bump commits.
-     * MUTUALLY EXCLUSIVE with is_first and is_last (application enforced).
-     */
-    isReleaseChore: integer('is_release_chore').notNull().default(0),
-  },
-  (table) => [
-    primaryKey({ columns: [table.releaseId, table.commitSha] }),
-    index('idx_release_commits_release_id').on(table.releaseId),
-    index('idx_release_commits_commit_sha').on(table.commitSha),
-    index('idx_release_commits_position').on(table.releaseId, table.position),
-  ],
-);
-
-/**
- * `release_changes` — Editorial CHANGELOG generation layer (SPEC-T9345 §3.7).
- *
- * Each row represents one bullet in the rendered CHANGELOG for a release.
- * A release with 10 features + 3 bugfixes + 1 breaking change has 14 rows here.
- *
- * `change_type` uses the CLEO 12-value taxonomy ({@link RELEASE_CHANGE_TYPES}) rather
- * than `tasks.kind`, so that:
- *   - A single task can appear in multiple releases with different change types
- *     (feature in v1, hotfix in v2, reverted in v3).
- *   - `hotfix` is a release-packaging decision, not a task property.
- *   - Classification is agent-writable without touching OWNER-WRITE-ONLY fields.
- *
- * `task_id` is nullable: some changes are not task-linked (e.g., a dependency bump
- * detected from commit diff analysis with no associated CLEO task).
- *
- * `summary` is limited to 200 characters (application-layer enforced; parity tests
- * validate the constraint is documented on the column).
- *
- * `id` is generated as a UUID-ish string via `$defaultFn`.
- *
- * FKs:
- *   release_id → releases(id) ON DELETE CASCADE
- *   task_id    → tasks(id)    ON DELETE SET NULL
- *
- * @task T9508
- * @epic T9491
- * @see SPEC-T9345 §3.7
- */
-export const releaseChanges = sqliteTable(
-  'release_changes',
-  {
-    /** UUID primary key generated at insert time. */
-    id: text('id')
-      .primaryKey()
-      .$defaultFn(() => crypto.randomUUID()),
-    /** FK → releases.id. ON DELETE CASCADE. */
-    releaseId: text('release_id')
-      .notNull()
-      .references(() => releases.id, { onDelete: 'cascade' }),
-    /**
-     * FK → tasks.id (ON DELETE SET NULL). Nullable — some changes are
-     * not task-linked (e.g., automated dependency bumps).
-     */
-    taskId: text('task_id').references(() => tasks.id, { onDelete: 'set null' }),
-    /**
-     * CLEO 12-value change taxonomy. See {@link RELEASE_CHANGE_TYPES}.
-     * Orthogonal to `tasks.kind` — classification happens at the release level.
-     */
-    changeType: text('change_type', { enum: RELEASE_CHANGE_TYPES }).notNull(),
-    /**
-     * User-facing one-liner summary for the CHANGELOG. MUST be ≤ 200 characters.
-     * Enforced at application layer; parity tests document this constraint.
-     */
-    summary: text('summary').notNull(),
-    /** Optional markdown body with additional detail (multi-line allowed). */
-    description: text('description'),
-    /**
-     * Semver bump impact assessment. See {@link RELEASE_IMPACTS}.
-     * Defaults to 'patch' (conservative — can be promoted to 'minor' or 'major').
-     */
-    impact: text('impact', { enum: RELEASE_IMPACTS }).notNull().default('patch'),
-    /**
-     * Provenance of the classification. See {@link RELEASE_CLASSIFIED_BY}.
-     * Starts as 'auto' on machine classification; updated to 'manual' or 'approved'
-     * when a human reviews.
-     */
-    classifiedBy: text('classified_by', { enum: RELEASE_CLASSIFIED_BY }).notNull().default('auto'),
-    /** ISO-8601 timestamp when this change was classified. */
-    classifiedAt: text('classified_at').notNull().default(sql`(datetime('now'))`),
-  },
-  (table) => [
-    index('idx_release_changes_release_id').on(table.releaseId),
-    index('idx_release_changes_task_id').on(table.taskId),
-    index('idx_release_changes_change_type').on(table.changeType),
-    index('idx_release_changes_impact').on(table.impact),
-  ],
-);
-
-/**
- * `release_changesets` — Persistence layer for CLEO-native task-anchored
- * changesets (T9738 carryforward → T9753).
- *
- * Each row records one `.changeset/*.md` entry that was rolled into the
- * corresponding release plan by `cleo release plan`. The plan verb parses
- * `.changeset/*.md` via {@link parseChangesetDir} (in
- * `@cleocode/core/changesets`), persists one row per entry here, and then
- * aggregates them into the CHANGELOG markdown section embedded into the
- * release plan envelope.
- *
- * Differs from {@link releaseChanges} (T9508): `releaseChanges` is the
- * editorial CHANGELOG bucket derived from commit-level analysis — one row per
- * rendered bullet. `releaseChangesets` is the upstream-entry source — one row
- * per `.changeset/*.md` file, carrying the structured frontmatter as-is so
- * downstream renderers can re-aggregate without re-reading the markdown.
- *
- * `taskIds` / `prs` are JSON-encoded arrays. `notes` and `breaking` are
- * nullable longer-form bodies preserved verbatim from the source entry.
- *
- * FK: `release_id → releases(id) ON DELETE CASCADE` — purging a release
- * removes its changeset rows.
- *
- * @task T9753
- * @epic T9752
- */
-export const releaseChangesets = sqliteTable(
-  'release_changesets',
-  {
-    /** UUID primary key generated at insert time. */
-    id: text('id')
-      .primaryKey()
-      .$defaultFn(() => crypto.randomUUID()),
-    /** FK → releases.id. ON DELETE CASCADE — purging a release removes its changeset rows. */
-    releaseId: text('release_id')
-      .notNull()
-      .references(() => releases.id, { onDelete: 'cascade' }),
-    /** Filename slug of the `.changeset/<slug>.md` file (matches `ChangesetEntry.id`). */
-    changesetId: text('changeset_id').notNull(),
-    /** JSON array of CLEO task IDs anchored by this changeset entry. */
-    taskIds: text('task_ids').notNull(),
-    /** Kind discriminator — mirrors {@link CHANGESET_KINDS} from `@cleocode/contracts`. */
-    kind: text('kind').notNull(),
-    /** User-facing one-liner summary lifted verbatim from the entry's `summary` field. */
-    summary: text('summary').notNull(),
-    /** JSON array of integer PR numbers, nullable. */
-    prs: text('prs'),
-    /** Markdown body — longer-form explanation from the source entry, nullable. */
-    notes: text('notes'),
-    /** Migration note when `kind = 'breaking'`, nullable. */
-    breaking: text('breaking'),
-    /** ISO-8601 timestamp when this row was inserted. */
-    createdAt: text('created_at').notNull().default(sql`(datetime('now'))`),
-  },
-  (table) => [
-    index('release_changesets_release_id_idx').on(table.releaseId),
-    index('release_changesets_changeset_id_idx').on(table.changesetId),
-    index('release_changesets_kind_idx').on(table.kind),
-  ],
-);
-
-/**
- * Artifact type enum for {@link releaseArtifacts.artifactType}.
- *
- * Polymorphic: a single table covers all publishing archetypes. Adding a new
- * artifact type (e.g. `gem`, `maven`) requires adding a value here and writing
- * new rows — zero schema changes required.
- *
- * - `npm`            — Node.js package published to npmjs.com or a private registry
- * - `cargo`          — Rust crate published to crates.io
- * - `docker`         — Container image pushed to an OCI registry
- * - `pypi`           — Python package published to pypi.org
- * - `github-release` — GitHub Releases asset attached to a git tag
- * - `binary`         — Generic compiled binary distributed via direct URL
- * - `github-tag`     — Lightweight git tag (no attached assets)
- *
- * @task T9509
- * @see SPEC-T9345 §3.9
- */
-export const RELEASE_ARTIFACT_TYPES = [
-  'npm',
-  'cargo',
-  'docker',
-  'pypi',
-  'github-release',
-  'binary',
-  'github-tag',
-] as const;
-
-/**
- * Union type for {@link RELEASE_ARTIFACT_TYPES}. Promoted to
- * `@cleocode/contracts/provenance` in Phase 0c (T9955); re-exported here
- * for backward compatibility.
- */
-export type { ReleaseArtifactType };
-
-/**
- * Link type enum for {@link brainReleaseLinks.linkType}.
- *
- * Describes the semantic relationship between a BRAIN entry and a release:
- * - `approved-by`    — A BRAIN decision approved a change that shipped in this release.
- * - `documented-in`  — This release is where the BRAIN entry was first formally documented.
- * - `derived-from`   — The release's failure or outcome produced this BRAIN learning/pattern.
- * - `observed-in`    — A BRAIN observation was made about this release (e.g. performance note).
- *
- * @task T9509
- * @see SPEC-T9345 §8.1
- */
-export const BRAIN_RELEASE_LINK_TYPES = [
-  'approved-by',
-  'documented-in',
-  'derived-from',
-  'observed-in',
-] as const;
-
-/**
- * Union type for {@link BRAIN_RELEASE_LINK_TYPES}. Promoted to
- * `@cleocode/contracts/provenance` in Phase 0c (T9955); re-exported here
- * for backward compatibility.
- */
-export type { BrainReleaseLinkType };
-
-/**
- * `release_artifacts` — Polymorphic artifact registry (ADR-073 / SPEC-T9345 §3.9).
- *
- * One row per (release × artifact type × identifier) triple. A monorepo release
- * that publishes 22 npm packages + 1 cargo crate + 1 docker image produces 24 rows
- * for that single release.
- *
- * **Polymorphic design**: The `metadata` column is a JSON blob that holds
- * type-specific extras (e.g. `{"dist-tag":"latest","integrity":"sha512-..."}` for
- * npm; `{"digest":"sha256:abc","registry":"ghcr.io"}` for docker). This eliminates
- * the need for separate per-archetype tables and means adding a new artifact type
- * requires zero schema changes.
- *
- * `identifier` is type-specific:
- *   - npm      → package name (e.g. `@cleocode/cleo`)
- *   - cargo    → crate name (e.g. `cleo-core`)
- *   - docker   → image reference (e.g. `ghcr.io/org/cleo:v2026.6.0`)
- *   - binary   → filename or slug (e.g. `cleo-linux-x64`)
- *   - github-release / github-tag → tag name (e.g. `v2026.6.0`)
- *
- * Composite PRIMARY KEY: (release_id, artifact_type, identifier).
- *
- * FK:
- *   release_id → releases(id) ON DELETE CASCADE
- *
- * @task T9509
- * @epic T9491
- * @see SPEC-T9345 §3.9
- */
-export const releaseArtifacts = sqliteTable(
-  'release_artifacts',
-  {
-    /** FK → releases.id. ON DELETE CASCADE — purging a release removes its artifact rows. */
-    releaseId: text('release_id')
-      .notNull()
-      .references(() => releases.id, { onDelete: 'cascade' }),
-    /**
-     * Artifact archetype. See {@link RELEASE_ARTIFACT_TYPES}.
-     * Stored as plain text for forward-compat: adding a new type is a code-only change.
-     */
-    artifactType: text('artifact_type').notNull(),
-    /**
-     * Artifact-specific identifier.
-     * npm: package name; cargo: crate name; docker: image ref; binary: filename.
-     */
-    identifier: text('identifier').notNull(),
-    /** Published version string (artifact-specific format). */
-    version: text('version').notNull(),
-    /** Registry URL, OCI ref, or asset download URL. Nullable (not all archetypes have one). */
-    url: text('url'),
-    /** ISO-8601 timestamp when the artifact was published to its registry. */
-    publishedAt: text('published_at'),
-    /**
-     * JSON blob for type-specific metadata.
-     * npm: `{"dist-tag":"latest","integrity":"sha512-...","provenance":true}`
-     * docker: `{"digest":"sha256:abc","registry":"ghcr.io"}`
-     * cargo: `{"checksum":"sha256:def","yanked":false}`
-     * Defaults to `{}` for archetypes that have no extras.
-     */
-    metadata: text('metadata').notNull().default('{}'),
-  },
-  (table) => [
-    primaryKey({ columns: [table.releaseId, table.artifactType, table.identifier] }),
-    index('idx_release_artifacts_release_id').on(table.releaseId),
-    index('idx_release_artifacts_artifact_type').on(table.artifactType),
-    index('idx_release_artifacts_published_at').on(table.publishedAt),
-  ],
-);
-
-/**
- * `brain_release_links` — M:N junction closing the BRAIN↔release loop
- * (ADR-073 / SPEC-T9345 §8).
- *
- * Links BRAIN entries (decisions, observations, patterns, learnings stored in
- * the project-level `.cleo/brain.db`) to releases stored in `tasks.db`, so that
- * queries like "Which decision approved the fix in v2026.5.74?" are SQL-answerable.
- *
- * **Cross-DB soft FK strategy** (SPEC-T9345 §8.2):
- * `brain_entry_id` is an intentionally un-constrained TEXT column. `brain.db` is
- * a separate SQLite file and SQLite cannot enforce cross-file foreign keys.
- * `ON DELETE SET NULL` semantics are handled at the application layer — when a
- * BRAIN entry is deleted, the calling code sets `brain_entry_id = NULL` on
- * associated rows (or removes the link row entirely). This matches the existing
- * cross-DB pattern in `packages/core/src/store/brain-accessor-impl.ts`.
- *
- * `release_id` carries a hard REFERENCES constraint to `releases(id)` in the
- * same `tasks.db` file. ON DELETE CASCADE: purging a release removes all its
- * brain link rows.
- *
- * Composite PRIMARY KEY: (brain_entry_id, release_id, link_type).
- * Multiple link_type values between the same pair are valid — a decision can
- * both `approved-by` and `documented-in` the same release.
- *
- * @task T9509
- * @epic T9491
- * @see SPEC-T9345 §8.1
- */
-export const brainReleaseLinks = sqliteTable(
-  'brain_release_links',
-  {
-    /**
-     * Soft FK to `brain_entries.id` in `brain.db`.
-     * NOT a hard REFERENCES — brain.db is a separate SQLite file.
-     * Nullable: application sets to NULL when the BRAIN entry is deleted (SET NULL semantics).
-     */
-    brainEntryId: text('brain_entry_id'),
-    /** FK → releases.id. ON DELETE CASCADE. */
-    releaseId: text('release_id')
-      .notNull()
-      .references(() => releases.id, { onDelete: 'cascade' }),
-    /**
-     * Semantic relationship type. See {@link BRAIN_RELEASE_LINK_TYPES}.
-     * - `approved-by`   — BRAIN decision approved this release's change.
-     * - `documented-in` — Release where the BRAIN entry was first documented.
-     * - `derived-from`  — Release outcome produced this BRAIN learning/pattern.
-     * - `observed-in`   — BRAIN observation made about this release.
-     */
-    linkType: text('link_type', { enum: BRAIN_RELEASE_LINK_TYPES }).notNull(),
-    /** ISO-8601 timestamp when this link was created. */
-    createdAt: text('created_at').notNull().default(sql`(datetime('now'))`),
-    /** Identifier of the agent or user that created this link. Nullable. */
-    createdBy: text('created_by'),
-  },
-  (table) => [
-    primaryKey({ columns: [table.brainEntryId, table.releaseId, table.linkType] }),
-    index('idx_brain_release_links_brain_entry_id').on(table.brainEntryId),
-    index('idx_brain_release_links_release_id').on(table.releaseId),
-    index('idx_brain_release_links_link_type').on(table.linkType),
-  ],
-);
-
-// === TYPE EXPORTS ===
-
-export type AttachmentRow = typeof attachments.$inferSelect;
-export type NewAttachmentRow = typeof attachments.$inferInsert;
-export type AttachmentRefRow = typeof attachmentRefs.$inferSelect;
-export type NewAttachmentRefRow = typeof attachmentRefs.$inferInsert;
-export type TaskRow = typeof tasks.$inferSelect;
-export type NewTaskRow = typeof tasks.$inferInsert;
-export type SessionRow = typeof sessions.$inferSelect;
-export type NewSessionRow = typeof sessions.$inferInsert;
-export type TaskDependencyRow = typeof taskDependencies.$inferSelect;
-export type TaskRelationRow = typeof taskRelations.$inferSelect;
-export type WorkHistoryRow = typeof taskWorkHistory.$inferSelect;
-export type LifecyclePipelineRow = typeof lifecyclePipelines.$inferSelect;
-export type NewLifecyclePipelineRow = typeof lifecyclePipelines.$inferInsert;
-export type LifecycleStageRow = typeof lifecycleStages.$inferSelect;
-export type NewLifecycleStageRow = typeof lifecycleStages.$inferInsert;
-export type LifecycleGateResultRow = typeof lifecycleGateResults.$inferSelect;
-export type NewLifecycleGateResultRow = typeof lifecycleGateResults.$inferInsert;
-export type LifecycleEvidenceRow = typeof lifecycleEvidence.$inferSelect;
-export type NewLifecycleEvidenceRow = typeof lifecycleEvidence.$inferInsert;
-export type LifecycleTransitionRow = typeof lifecycleTransitions.$inferSelect;
-export type NewLifecycleTransitionRow = typeof lifecycleTransitions.$inferInsert;
-export type AuditLogRow = typeof auditLog.$inferSelect;
-export type NewAuditLogRow = typeof auditLog.$inferInsert;
-export type TokenUsageRow = typeof tokenUsage.$inferSelect;
-export type NewTokenUsageRow = typeof tokenUsage.$inferInsert;
-export type ArchitectureDecisionRow = typeof architectureDecisions.$inferSelect;
-export type NewArchitectureDecisionRow = typeof architectureDecisions.$inferInsert;
-export type AdrTaskLinkRow = typeof adrTaskLinks.$inferSelect;
-export type NewAdrTaskLinkRow = typeof adrTaskLinks.$inferInsert;
-export type AdrRelationRow = typeof adrRelations.$inferSelect;
-export type NewAdrRelationRow = typeof adrRelations.$inferInsert;
-export type ManifestEntryRow = typeof manifestEntries.$inferSelect;
-export type NewManifestEntryRow = typeof manifestEntries.$inferInsert;
-export type PipelineManifestRow = typeof pipelineManifest.$inferSelect;
-export type NewPipelineManifestRow = typeof pipelineManifest.$inferInsert;
-// `ReleaseManifestRow` / `NewReleaseManifestRow` removed by T9686-B2.
-// Use `ReleaseRow` / `NewReleaseRow` (now backed by the unified `releases`
-// table) instead. Legacy-only columns (`tasksJson`, `changelog`, `notes`,
-// `gitTag`, `preparedAt`, `committedAt`, `taggedAt`, `pushedAt`) are
-// nullable fields on the unified row type.
-export type ExternalTaskLinkRow = typeof externalTaskLinks.$inferSelect;
-export type NewExternalTaskLinkRow = typeof externalTaskLinks.$inferInsert;
-// T944 experiments side-table row types
-export type ExperimentRow = typeof experiments.$inferSelect;
-export type NewExperimentRow = typeof experiments.$inferInsert;
-// T9506 provenance graph row types
-export type CommitRow = typeof commits.$inferSelect;
-export type NewCommitRow = typeof commits.$inferInsert;
-export type TaskCommitRow = typeof taskCommits.$inferSelect;
-export type NewTaskCommitRow = typeof taskCommits.$inferInsert;
-export type CommitFileRow = typeof commitFiles.$inferSelect;
-export type NewCommitFileRow = typeof commitFiles.$inferInsert;
-// T9507 PR provenance graph row types
-export type PullRequestRow = typeof pullRequests.$inferSelect;
-export type NewPullRequestRow = typeof pullRequests.$inferInsert;
-export type PrCommitRow = typeof prCommits.$inferSelect;
-export type NewPrCommitRow = typeof prCommits.$inferInsert;
-export type PrTaskRow = typeof prTasks.$inferSelect;
-export type NewPrTaskRow = typeof prTasks.$inferInsert;
-// T9508 / T9686-B2: unified releases table row types
-export type ReleaseRow = typeof releases.$inferSelect;
-export type NewReleaseRow = typeof releases.$inferInsert;
-export type ReleaseCommitRow = typeof releaseCommits.$inferSelect;
-export type NewReleaseCommitRow = typeof releaseCommits.$inferInsert;
-export type ReleaseChangeRow = typeof releaseChanges.$inferSelect;
-export type NewReleaseChangeRow = typeof releaseChanges.$inferInsert;
-// T9753 release changesets (CLEO-native changesets aggregator)
-export type ReleaseChangesetRow = typeof releaseChangesets.$inferSelect;
-export type NewReleaseChangesetRow = typeof releaseChangesets.$inferInsert;
-// T9509 provenance graph row types (release_artifacts + brain_release_links)
-export type ReleaseArtifactRow = typeof releaseArtifacts.$inferSelect;
-export type NewReleaseArtifactRow = typeof releaseArtifacts.$inferInsert;
-export type BrainReleaseLinkRow = typeof brainReleaseLinks.$inferSelect;
-export type NewBrainReleaseLinkRow = typeof brainReleaseLinks.$inferInsert;
-
-// agent_credentials REMOVED from tasks.db — T234 clean-cut.
-// Agent data (identity, credentials, capabilities, skills) now lives
-// exclusively in signaldock.db. See AgentRegistryAccessor.
